% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balance_inds.R
\name{balance_inds}
\alias{balance_inds}
\title{Split a sequence of integers into a partition with roughly equal-sized parts}
\usage{
balance_inds(n.inds, n.partitions)
}
\arguments{
\item{n.inds}{Defines the number range \code{1:n.inds} that will be split
into roughly equal-sized parts.}

\item{n.partitions}{The number of groups to create.}
}
\description{
Group the numbers \code{1:n.inds} into \code{n.partitions} such that all of 
the groups are about the same size.  This method can be useful for batching 
a complex operation into roughly equal-sized parts that will be evaluated in 
parallel.
}
\examples{
balance_inds(n.inds = 1e3, n.partitions = 7)
balance_inds(n.inds = 3, n.partitions = 7)
balance_inds(n.inds = 8, n.partitions = 7)
}
