% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crawl.impute.R
\name{crawl.impute}
\alias{crawl.impute}
\title{Use bridged sampling to impute a dive trajectory consistent with observations}
\usage{
crawl.impute(depth.bins, depths, times, N, depths.impute = "uniform", tstep)
}
\arguments{
\item{depth.bins}{\eqn{n x 2} Matrix that defines the depth bins.  The first 
column defines the depth at the center of each depth bin, and the second 
column defines the half-width of each bin.}

\item{depths}{record of depth bins the trajectory should visit}

\item{times}{times at which the depth bins should be visited}

\item{N}{number of trajectories to impute}

\item{depths.impute}{Either \code{'uniform'} or \code{'midpoint'}.  The 
\code{crawl} model requires continuous rather than discrete depth values.
Setting \code{depths.impute='uniform'} will fit the \code{crawl} model 
using depth values that are uniformly sampled from each depth bin.  To 
average over the uniform imputation, each of the \code{N} imputed 
trajectories will fit the \code{crawl} model to a different set of 
uniformly sampled depths.
Setting \code{depths.impute='midpoint'} will fith the \code{crawl} model 
using the midpoint of each depth bin.}

\item{tstep}{Time step to be used when imputing continuous trajectories to be 
discretized}

\item{s0}{dive stage at which the trajectory should be started from}

\item{beta}{\eqn{2 x 3} matrix in which each column contains the diving 
preference and directional persistence parameters for the DIVING, SUBMERGED, 
and SURFACING dive stages.}

\item{lambda}{length 3 vector that specifies the transition rate, 
respectively in the DIVING, SUBMERGED, and SURFACING stages.}

\item{sub.tx}{length 2 vector that specifies the first depth bin at which 
transitions to the SUBMERGED stage can occur and the probability that such 
a transition occurs at the next depth transition}

\item{surf.tx}{parameter that specifies the probability the trajectory will 
transition to the SURFACING stage at the next depth transition}

\item{inflation.factor.lambda}{In order to facilitate bridged transitions, 
the transition rate of the overall process must be inflated to allow the 
possibility of self-transitions.  Self-transitions allow bridged paths to 
dynamically modify the total number of transitions between observed values
so that a valid path between observations is always possible.  The 
\code{inflation.factor.lambda} parameter implicitly controls the number of 
self-transitions that will occur.  Larger values will create more 
self-transitions.}

\item{verbose}{If \code{TRUE}, then the sampler's progress will be printed 
during sampling.}

\item{precompute.bridges}{If \code{TRUE}, then the bridged transition 
matrices will be precomputed.  Enabling this option will increase the 
memory overhead of the method, but will reduce its runtime.}
}
\description{
The imputation will be done via crawl only, so the imputation only imputes 
likely locations in discrete space.  Dive stages will not be imputed.  For 
inference, the dive stages should be treated as missing data.
}
\examples{
data('dive.sim')
attach(dive.sim)

imputed = crawl.impute(depth.bins = depth.bins, depths = sim.obs$depths, 
                       times = sim.obs$times, N = 1, depths.impute = 'midpoint', 
                       tstep = 5)

detach(dive.sim)
}
