% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dive.sim}
\alias{dive.sim}
\title{Simulated three stage dive}
\format{An object of class \code{list} of length 4.}
\usage{
dive.sim
}
\description{
The \code{dive.sim} object contains the complete simulated dive trajectory 
and as a corresponding set of observations in addition to the model 
parameters used to generate the dive.
}
\examples{

data("dive.sim")
str(dive.sim)

}
\keyword{datasets}
