% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dN.bridged.R
\name{dN.bridged}
\alias{dN.bridged}
\title{Use bridged sampling to impute a complete dive trajectory consistent with observations}
\usage{
dN.bridged(B, x0, xN, N.max, rate.uniformized, t, log = FALSE)
}
\arguments{
\item{B}{single-step transition matrix, after uniformization}

\item{x0}{state at which chain begins}

\item{xN}{state at which chain ends}

\item{rate.uniformized}{uniformized poisson process transition rate}

\item{t}{time period between x0 and xN observations}

\item{N}{maximum number of steps to consider}
}
\description{
The sampling method is designed to sample many trajectories simultaneously, 
so has an extra level of approximation in the proposal distributions.
}
\examples{
data('dive.sim')
attach(dive.sim)
attach(dive.sim$params)

r.unif = max(outer(lambda, 2 * depth.bins$halfwidth, '/'))

m = dsdive.tx.matrix.uniformized(depth.bins = depth.bins, beta = beta, 
                                 lambda = lambda, s0 = 2, 
                                 rate.uniformized = r.unif)

N.max = 20

dN = dN.bridged(B = m, x0 = 3, xN = 5, N.max = N.max, rate.uniformized = r.unif, 
                t = 60, log = FALSE)

plot(0:N.max, dN)

detach(dive.sim$params)
detach(dive.sim)
}
