% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.df.R
\name{ds.df}
\alias{ds.df}
\title{Builds a tidy data frame for observations of a dive}
\usage{
ds.df(depths, times, depth.bins, stages = NULL)
}
\arguments{
\item{depths}{Record of observed depth bins}

\item{times}{Times (in seconds) at which depth bin observations were made}

\item{depth.bins}{Vector that defines the depth bins (i.e., bin labels)}

\item{stages}{Vector of guesses for which dive stage the trajectory was 
in at each observation}
}
\description{
Converts raw dive information to a tidy dataframe that
is suitable for plotting with \code{ggplot2}.
}
\examples{
data('dive.sim')
attach(dive.sim)

df = ds.df(depths = sim.obs$depths, times = sim.obs$times, 
           depth.bins = depth.bins, stages = NULL)
}
