% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsdive.align.obs.R
\name{dsdive.align.obs}
\alias{dsdive.align.obs}
\title{Realign observations of a dive to the time scale of the dive}
\usage{
dsdive.align.obs(depths, times, t.stages, offset, offset.tf)
}
\arguments{
\item{depths}{Indices of observed depth bins}

\item{times}{Times at which each of \code{depths} was observed}

\item{t.stages}{Stage transition times for the dive; will be used to compute
the dive stage for each observation}

\item{offset}{Amount of time by which \code{times} over or under reports the 
true time within the dive of each \code{depths} observation}

\item{offset.tf}{Amount of time by which the last \code{time} over or under 
reports the true duration of the dive.}
}
\description{
Satellite tag observations of a dive record depths at regular intervals, 
irrespective of whether or not an animal is diving.  Given initial and final 
offsets \code{offset} and \code{offset.tf}, respectively, this function will 
shift the satellite-tag observation \code{times} so that they are referenced 
with respect to the start of the dive (i.e., t=0).  Observations in 
\code{depths} will be trimmed from the output if they exceed the times 
implied by the offsets.
}
\examples{
data('dive.sim')
attach(dive.sim)
attach(params)

t.stages = sim$times[c(FALSE, diff(sim$stages)==1)]
  
dsdive.align.obs(depths = sim.obs$depths, times = sim.obs$times, 
                 t.stages = t.stages, offset = -300, 
                 offset.tf = -300)

detach(params)
detach(dive.sim)
}
