% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsdive.augment.trajectory.R
\name{dsdive.augment.trajectory}
\alias{dsdive.augment.trajectory}
\title{Compute durations and stages from a complete record of depths and times}
\usage{
dsdive.augment.trajectory(depths, times, t.stages)
}
\description{
This function associates durations and stage labels with a complete 
observation record of all depth bin transitions, and stage transition times.
}
\details{
In practice, this is a support function that will facilitate re-computing 
durations given new stage transition times.
}
\examples{
data('dive.sim')
attach(dive.sim)

# recompute dive stages and durations when we change t.stages
augmented = dsdive.augment.trajectory(depths = sim$depths, times = sim$times, 
                                      t.stages = 60*c(20,40))

# dive before changing t.stages
pl1 = plot(x = sim, depth.bins = depth.bins)

# dive after changing t.stages
pl2 = plot(x = augmented, depth.bins = depth.bins)

detach(dive.sim)
 
}
