% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsdive.bridgesample.R
\name{dsdive.bridgesample}
\alias{dsdive.bridgesample}
\title{Simulate dive trajectories across discrete depth bins}
\usage{
dsdive.bridgesample(depth.bins, d0, d0.last, df, beta, lambda, sub.tx,
  surf.tx, t0, tf, s0, inflation.factor.lambda = 1.1, verbose = FALSE,
  precompute.bridges = FALSE)
}
\arguments{
\item{depth.bins}{\eqn{n x 2} Matrix that defines the depth bins.  The first 
column defines the depth at the center of each depth bin, and the second 
column defines the half-width of each bin.}

\item{d0}{the depth bin at which transition parameters should be computed}

\item{beta}{\eqn{2 x 3} matrix in which each column contains the diving 
preference and directional persistence parameters for the DIVING, SUBMERGED, 
and SURFACING dive stages.}

\item{lambda}{length 3 vector that specifies the transition rate, 
respectively in the DIVING, SUBMERGED, and SURFACING stages.}

\item{sub.tx}{length 2 vector that specifies the first depth bin at which 
transitions to the SUBMERGED stage can occur and the probability that such 
a transition occurs at the next depth transition}

\item{surf.tx}{parameter that specifies the probability the trajectory will 
transition to the SURFACING stage at the next depth transition}

\item{t0}{time at which transition parameters should be computed}

\item{tf}{time at which sampling should end after}

\item{s0}{dive stage in which forward simulation begins}
}
\value{
A \code{dsdive} object, which is a \code{list} with the following 
  vectors:
  \describe{
    \item{depths}{Record of which depth bin indices the trajectory visited}
    \item{durations}{Record of amount of time spent in each depth bin}
    \item{times}{The time at which each depth bin was entered}
    \item{stages}{The stage at which each depth bin was entered}
  }
}
\description{
The method will simulate dive trajectories from initial conditions until the 
trajectory is observable at \code{tf}, or a maximum number of transitions 
has been exceeded.  The dive simulation is bridged, so the trajectory will
also stop diving after returning to the surface.
}
\examples{
data('dive.sim')
attach(dive.sim)
attach(params)

br = dsdive.bridgesample(depth.bins = depth.bins, d0 = sim.obs$depths[1], 
                         d0.last = NULL, df = sim.obs$depths[2], beta = beta, 
                         lambda = lambda, sub.tx = sub.tx, surf.tx = surf.tx, 
                         t0 = sim.obs$times[1], tf = sim.obs$times[2], s0 = 1,
                         verbose = TRUE, precompute.bridges = TRUE)
}
