% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsdive.extract.R
\name{dsdive.extract}
\alias{dsdive.extract}
\title{Extract a segment of a fully observed dive trajectory between timepoints}
\usage{
dsdive.extract(depths, times, stages, durations, t0, tf)
}
\arguments{
\item{depths}{Complete record of depth bin indices visited}

\item{times}{Times at which each of \code{depths} was visited}

\item{stages}{Stages at which each of the \code{depths} was visited}

\item{durations}{the amount of time spent in each depth bin}

\item{t0}{the time to mark the beginning of the segment to extract}

\item{tf}{the time to mark the end of the segment to extract

# @example examples/dsdive.extract.R}
}
\description{
Extract a segment of a fully observed dive trajectory between timepoints
}
