% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsdive.fwdsample.dive.R
\name{dsdive.fwdsample.dive}
\alias{dsdive.fwdsample.dive}
\title{Simulate dive trajectories across discrete depth bins}
\usage{
dsdive.fwdsample.dive(depth.bins, beta, lambda, t0, steps.max, T1, T2)
}
\arguments{
\item{depth.bins}{\eqn{n x 2} Matrix that defines the depth bins.  The first 
column defines the depth at the center of each depth bin, and the second 
column defines the half-width of each bin.}

\item{beta}{\eqn{2 x 3} matrix in which each column contains the diving 
preference and directional persistence parameters for the DIVING, SUBMERGED, 
and SURFACING dive stages.}

\item{lambda}{length 3 vector that specifies the transition rate, 
respectively in the DIVING, SUBMERGED, and SURFACING stages.}

\item{t0}{time at which transition parameters should be computed}

\item{steps.max}{maximum number of transitions to sample before stopping, 
regardless of whether \code{tf} is reached.}

\item{T1}{stage 1 length (in same units as t0, d0, tf)}

\item{T2}{stage 2 length (in same units as t0, d0, tf)}

\item{d0}{the depth bin at which transition parameters should be computed}

\item{sub.tx}{length 2 vector that specifies the first depth bin at which 
transitions to the SUBMERGED stage can occur and the probability that such 
a transition occurs at the next depth transition}

\item{surf.tx}{parameter that specifies the probability the trajectory will 
transition to the SURFACING stage at the next depth transition}

\item{tf}{time at which sampling should end after}

\item{dur0}{time spent at location \code{d0}.  If \code{NULL}, then a 
duration in state \code{d0} will be sampled, otherwise a new state will 
be sampled first, then sampling will continue from the new state at time 
\code{t0 + dur0}.}

\item{nsteps}{if \code{nsteps} is not \code{NULL}, then the sampler will
be reconfigured to sample exactly \code{nsteps} transitions instead of 
attempting to sample until the trajectory is observable at time \code{tf}.}

\item{s0}{dive stage in which forward simulation begins}

\item{t0.dive}{Time at which dive started}

\item{shift.params}{Optional arguments to bias sampling toward a specific 
node.  See documentation for \code{dsdive.tx.params} for more detail.}

\item{model}{Either \code{"conditional"} or \code{"logit"} depending on the 
method used to determine stage transition probability curves}
}
\value{
A \code{dsdive} object, which is a \code{list} with the following 
  vectors:
  \describe{
    \item{depths}{Record of which depth bin indices the trajectory visited}
    \item{durations}{Record of amount of time spent in each depth bin}
    \item{times}{The time at which each depth bin was entered}
    \item{stages}{The stage at which each depth bin was entered}
  }
}
\description{
The method will simulate dive trajectories from initial conditions until the 
trajectory is observable at \code{tf}, or a maximum number of transitions 
has been exceeded.  The dive simulation is bridged, so the trajectory will
also stop diving after returning to the surface.
}
\examples{
data('dive.sim')
attach(dive.sim)
attach(dive.sim$params)

# simulate dive
x = dsdive.fwdsample.dive(depth.bins = depth.bins, beta = beta, 
                          lambda = lambda, t0 = 0, steps.max = 1e3, 
                          T1 = 15*60, T2 = 12*60)

pl = plot(x = x, depth.bins = depth.bins)

detach(dive.sim$params)
detach(dive.sim)
}
