% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsdive.fwdsample.fixedstage.R
\name{dsdive.fwdsample.fixedstage}
\alias{dsdive.fwdsample.fixedstage}
\title{Simulate a dive trajectory within a single stage}
\usage{
dsdive.fwdsample.fixedstage(
  depth.bins,
  d0,
  beta,
  lambda,
  t0,
  tf,
  steps.max,
  dur0 = NULL,
  nsteps = NULL,
  s0
)
}
\arguments{
\item{depth.bins}{\eqn{n x 2} Matrix that defines the depth bins.  The first 
column defines the depth at the center of each depth bin, and the second 
column defines the half-width of each bin.}

\item{d0}{the depth bin index at which the dive segment begins}

\item{beta}{Directional preference model parameters.  See 
\code{dsdive.tx.params} for more details.}

\item{lambda}{Diving rate model parameters.  See 
\code{dsdive.tx.params} for more details.}

\item{t0}{Initial time for the dive segment}

\item{tf}{Final time for the dive segment}

\item{steps.max}{maximum number of transitions to sample before stopping, 
regardless of whether \code{tf} is reached.}

\item{dur0}{time spent at location \code{d0}.  If \code{NULL}, then a 
duration in state \code{d0} will be sampled first, otherwise a new state 
will  be sampled first, then sampling will continue from the new state at  
time \code{t0 + dur0}.}

\item{nsteps}{if \code{nsteps} is not \code{NULL}, then the sampler will
be reconfigured to sample exactly \code{nsteps} transitions instead of 
attempting to sample until the trajectory is observable at time \code{tf}.}

\item{s0}{dive stage in which forward simulation begins}
}
\value{
A \code{dsdive} object, which is a \code{list} with the following 
  vectors:
  \describe{
    \item{depths}{Record of which depth bin indices the trajectory visited}
    \item{durations}{Record of amount of time spent in each depth bin}
    \item{times}{The time at which each depth bin was entered}
    \item{stages}{The stage at which each depth bin was entered}
  }
}
\description{
The method will simulate dive trajectories from initial conditions until the 
trajectory is observable at \code{tf}, or a maximum number of transitions 
has been exceeded.  The dive segment will only be simulated using parameters 
for a single dive stage \code{s0}.
}
\examples{
data('dive.sim')
attach(dive.sim)
attach(dive.sim$params)

# simulate dive
x = dsdive.fwdsample.fixedstage(depth.bins = depth.bins, d0 = 1, 
                                beta = beta, lambda = lambda, t0 = 0, 
                                tf = 100, steps.max = 1e3, s0 = 1)

pl = plot(x = x, depth.bins = depth.bins)

detach(dive.sim$params)
detach(dive.sim)
}
