% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsdive.generator.matrix.uniformized.R
\name{dsdive.generator.matrix.uniformized}
\alias{dsdive.generator.matrix.uniformized}
\title{Compute infinitesimal generator matrix for a homogeneous, uniformized CTMC}
\usage{
dsdive.generator.matrix.uniformized(
  depth.bins,
  beta,
  lambda,
  s0,
  rate.uniformized
)
}
\arguments{
\item{depth.bins}{\eqn{n x 2} Matrix that defines the depth bins.  The first 
column defines the depth at the center of each depth bin, and the second 
column defines the half-width of each bin.}

\item{beta}{Directional preference model parameters.  See 
\code{dsdive.tx.params} for more details.}

\item{lambda}{Diving rate model parameters.  See 
\code{dsdive.tx.params} for more details.}

\item{s0}{dive stage for which matrix should be computed}

\item{rate.uniformized}{uniformization rate, for standardizing transition
rates between states}
}
\description{
Compute infinitesimal generator matrix for a uniformized Continuous time
Markov chain (CTMC).  See Rao and Teh (2013) for an applied definition and 
use of infinitesimal generator matrices.
}
\examples{
data('dive.sim')
attach(dive.sim)
attach(dive.sim$params)

r.unif = max(outer(lambda, 2 * depth.bins$halfwidth, '/'))

m = dsdive.generator.matrix.uniformized(
  depth.bins = depth.bins, beta = beta, lambda = lambda, s0 = 2, 
  rate.uniformized = r.unif)

detach(dive.sim$params)
detach(dive.sim)
}
\references{
Rao, Vinayak, and Yee Whye Teh. "Fast MCMC sampling for Markov 
  jump processes and extensions." The Journal of Machine Learning Research 
  14.1 (2013): 3295-3320.
}
