% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsdive.heurest.R
\name{dsdive.heurest}
\alias{dsdive.heurest}
\title{Heuristic estimate of model parameters given observations}
\usage{
dsdive.heurest(depths, times, stages.est, depth.bins, priors = rep(5,
  12), sub.tx1)
}
\arguments{
\item{depths}{Record of observed depth bins}

\item{times}{Times at which depth bin observations were made}

\item{stages.est}{Vector of guesses for which dive stage the trajectory was 
in at each observation}

\item{depth.bins}{Vector that defines the depth bins (i.e., bin labels)}

\item{priors}{Standard deviations for mean-0 normal priors on model 
parameters, after transformation to real line}

\item{sub.tx1}{Index of minimum depth bin at which transition to intermediate 
behaviors stage can occur.}
}
\description{
Builds heuristic estimates of dive model parameters, which is useful for 
initializing methods that optimize or explore parameter surfaces and 
posterior distributions.
}
\examples{
data('dive.sim')
attach(dive.sim)

stages.est = c(rep(1,4), rep(2,15), rep(3, 7))

est = dsdive.heurest(depths = sim.obs$depths, times = sim.obs$times, 
                     stages.est = stages.est, depth.bins = depth.bins, 
                     sub.tx1 = params$sub.tx[1])
}
