% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsdive.impute.sample_n.R
\name{dsdive.impute.sample_n}
\alias{dsdive.impute.sample_n}
\title{Sample number of transitions between observations for a CTMC}
\usage{
dsdive.impute.sample_n(
  d0,
  df,
  s0,
  sf,
  t0,
  tf,
  t.stages,
  rate.unif,
  P.raw,
  P.tx,
  n.bins,
  max.tx
)
}
\arguments{
\item{d0}{index of depth bin in which the CTMC segment starts}

\item{df}{index of depth bin in which the CTMC segment ends}

\item{s0}{dive stage in which the CTMC begins}

\item{sf}{dive estage in which the CTMC ends}

\item{t0}{time at which \code{d0} is observed}

\item{tf}{time at which \code{df} is observed}

\item{t.stages}{stage transition times for the dive}

\item{rate.unif}{uniformization rate, for standardizing transition
rates between states}

\item{P.raw}{list of continuous time probability transition matrices, and 
components.}

\item{P.tx}{list of discrete time probability transition matrices}

\item{n.bins}{number of rows in the \code{depths} matrix}

\item{max.tx}{maximum number of transitions between observations that will 
be allowed during imputation}
}
\description{
Uses properties of homogeneous Continuous time Markov Chains (CTMCs) to 
sample the number of transitions between known endpoints and times.
}
