% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsdive.ld.R
\name{dsdive.ld}
\alias{dsdive.ld}
\title{Likelihood for completely observed dive trajectories}
\usage{
dsdive.ld(depths, durations, times, stages, beta, lambda, sub.tx, surf.tx,
  depths.labels)
}
\description{
Likelihood for completely observed dive trajectories
}
\examples{
data('dive.sim')
attach(dive.sim)
attach(dive.sim$params)

ld = dsdive.ld(depths = sim$depths, durations = sim$durations, 
               times = sim$times, stages = sim$stages, beta = beta, 
               lambda = lambda, sub.tx = sub.tx, surf.tx = surf.tx, 
               depths.labels = depth.bins)
}
