% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsdive.obs.sample.offsets.R
\name{dsdive.obs.sample.offsets}
\alias{dsdive.obs.sample.offsets}
\title{Sample t0 and tf offsets from full conditional posteriors}
\usage{
dsdive.obs.sample.offsets(
  dsobs.aligned,
  dsobs.unaligned,
  offset,
  offset.tf,
  t.stages,
  P.raw,
  depth.bins,
  tstep,
  max.width,
  prior.params,
  sample.start,
  debug = FALSE,
  lpapprox = NULL,
  output.lpapprox = FALSE
)
}
\arguments{
\item{dsobs.aligned}{\code{dsobs} object that has been adjusted to account 
for the t0 and tf offsets.  Note that this should be based on the current 
value of the offsets, and that this function will update these parameters.}

\item{dsobs.unaligned}{\code{dsobs} object with the raw observations of a 
dive object}

\item{offset}{value of current t0 offset}

\item{offset.tf}{value of current tf offset}

\item{t.stages}{stage transition times for the dive}

\item{P.raw}{list of continuous time probability transition matrices, and 
components.}

\item{depth.bins}{\eqn{n x 2} Matrix that defines the depth bins.  The first 
column defines the depth at the center of each depth bin, and the second 
column defines the half-width of each bin.}

\item{tstep}{Time between observations in \code{dsobs.unaligned}}

\item{max.width}{The t0 and tf offsets are updated with a piecewise 
proposal distribution.  This parameter controls the maximum width 
of the intervals for the proposal distribution.  This is a tuning parameter 
that controls the numerical stability of the proposal distribution, which 
is sampled via inverse CDF techniques.}

\item{prior.params}{\code{shape1} and \code{shape2} parameters for the 
scaled and shifted Beta prior distribution for the offset being sampled.}

\item{sample.start}{\code{TRUE} to sample t0 offset; 
\code{FALSE} to sample tf offset.}

\item{debug}{\code{TRUE} to return debugging objects, such as the proposal 
density and log posterior}

\item{lpapprox}{(Optional) Object containing the piecewise-quadratic
approximation to the log-posterior used to propose model parameters.  
If \code{NULL}, then an approximation will be computed.}

\item{output.lpapprox}{\code{TRUE} to return the approximation used 
  to propose model parameters.
  
# @example examples/dsdive.obs.sample.offsets.R}
}
\description{
Sampler uses a piecewise quadratic polynomial approximation to the log of the 
full conditional posterior.  Note that this sampler will only sample one 
offset at a time.
}
