% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsdive.obs.sampleparams.R
\name{dsdive.obs.sampleparams}
\alias{dsdive.obs.sampleparams}
\title{Full conditional distribution for dive stages given data and parameters}
\usage{
dsdive.obs.sampleparams(
  dsobs.list,
  t.stages.list,
  P.raw,
  s0,
  depth.bins,
  beta,
  lambda,
  lambda.priors.list,
  beta.priors.list,
  tstep
)
}
\arguments{
\item{P.raw}{current collection of transition probability matrices}

\item{s0}{the stage for which updated parameters should be sampled}

\item{beta}{the current depth bin transition model parameters}

\item{lambda}{the current depth bin transition rate model parameters}
}
\description{
Full conditional distribution for dive stages given data and parameters
}
\examples{
data('dive.sim')
attach(dive.sim)
attach(dive.sim$params)

t.stages = sim$times[c(FALSE,diff(sim$stages)==1)]
  
tstep = diff(sim.obs$times[1:2])

obstx.mat = lapply(1:3, function(s) {
  dsdive.obstx.matrix(depth.bins = depth.bins, beta = beta, 
                      lambda = lambda, s0 = s, tstep = tstep, 
                      include.raw = TRUE)
})

lambda.priors = list(
  c(4, 2 + 2/3), c(4, 13 + 1/3), c(4, 5)
)

beta.priors = list(
  c(31.5, 3.5), c(3.5, 31.5)
)

x = dsdive.obs.sampleparams(dsobs.list = list(sim.obs), 
                            t.stages.list = list(t.stages), P.raw = obstx.mat, 
                            s0 = 1, depth.bins = depth.bins, beta = beta, 
                            lambda = lambda, lambda.priors.list = lambda.priors, 
                            beta.priors.list = beta.priors)

detach(dive.sim$params)
detach(dive.sim)
}
