% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsdive.observe.R
\name{dsdive.observe}
\alias{dsdive.observe}
\title{Simulate observing a fully observed dive trajectory at specified timepoints}
\usage{
dsdive.observe(depths, times, t.obs)
}
\arguments{
\item{depths}{Complete record of depth bins visited}

\item{times}{Times at which each of \code{depths} was visited}

\item{t.obs}{Times at which the trajectory should be observed}
}
\description{
Simulate observing a fully observed dive trajectory at specified timepoints
}
\examples{
data('dive.sim')
attach(dive.sim)

obs = dsdive.observe(depths = sim$depths, times = sim$times, 
                     t.obs = seq(from = 0, to = max(sim$times)+60, by = 60))
}
