% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsdive.obsld.R
\name{dsdive.obsld}
\alias{dsdive.obsld}
\title{Likelihood for collection of partially observed dives}
\usage{
dsdive.obsld(dsobs.list, t.stages.list, P.raw, s0, sf)
}
\arguments{
\item{dsobs.list}{list of \code{dsobs} objects, which describe the 
observation times and depths of a collection of dives}

\item{t.stages.list}{list of stage transition times for dives 
observed in \code{dsobs.list}}

\item{P.raw}{list of continuous time probability transition matrices, and 
components.}

\item{s0}{likelihood should include contributions for stages greater or 
equal to \code{s0}}

\item{sf}{likelihood should include contributions for stages less than or 
equal to \code{sf}}
}
\description{
Likelihood for collection of partially observed dives
}
\examples{
library(dsdive)
data('dive.sim')
attach(dive.sim)
attach(dive.sim$params)

# extract time between observations
tstep = diff(sim.obs$times[1:2])

# get true stage transition times
t.stages = sim$times[c(FALSE, diff(sim$stages)==1)]

# build probability matrices for observations
obstx.mat = lapply(1:3, function(s) {
  dsdive.obstx.matrix(depth.bins = depth.bins, beta = dive.sim$params$beta, 
                      lambda = lambda, 
                      s0 = s, tstep = tstep, include.raw = TRUE, delta = 1e-10)
})

# compute likelihood of observations, given model parameters
ld = dsdive.obsld(dsobs.list = list(sim.obs,sim.obs),
                  t.stages.list = list(t.stages,t.stages), 
                  P.raw = obstx.mat, s0 = 1, sf = 3)

detach(dive.sim$params)
detach(dive.sim)
}
