% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsdive.prior.R
\name{dsdive.prior}
\alias{dsdive.prior}
\title{Evaluates prior distribution at parameter values}
\usage{
dsdive.prior(par, spec, log = TRUE)
}
\arguments{
\item{par}{list specifying model parameters.  the list has entries: 
\code{beta}, \code{lambda}, \code{sub.tx}, \code{surf.tx}}

\item{spec}{list specifying parameters for prior distributions.  the list has 
entries:
\describe{
  \item{beta.sd}{standard deviations for truncated normal prior 
    distributions associated with beta.}
  \item{beta.absmax}{maximum of absolute values for transformation out of 
    normal prior distributions associated with beta.}
  \item{lambda.sd}{standard deviations for centered normal prior 
    distributions associated with lambda.}
  \item{sub.tx.mean}{means for normal prior distributions associated with 
    stage 1->2 transition.}
  \item{sub.tx.sd}{standard devations for normal prior distributions 
    associated with stage 1->2 transition.}
  \item{surf.tx.mean}{means for normal prior distributions associated with 
    stage 2->3 transition.}
  \item{surf.tx.sd}{standard devations for normal prior distributions 
    associated with stage 2->3 transition.}
}}

\item{log}{\code{TRUE} to return prior on log scale}
}
\description{
Note that the model parameters are converted to their unrestricted supports,
on which the prior distributions are defined.
}
\examples{
data('dive.sim')
attach(dive.sim)
attach(params)

spec = list(beta.sd = rep(1, 3), beta.absmax = 10, lambda.sd = 1, 
            sub.tx.mean = 0, sub.tx.sd = 1, surf.tx.mean = 0, surf.tx.sd = 1)

dsdive.prior(par = params, spec = spec)

detach(params)
detach(dive.sim)
}
