% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsdive.sample.stages.R
\name{dsdive.sample.stages}
\alias{dsdive.sample.stages}
\title{Sample stage transition times from full conditional posterior}
\usage{
dsdive.sample.stages(
  depths,
  durations,
  times,
  t.stages,
  beta,
  lambda,
  depth.bins,
  T1.prior,
  T2.prior
)
}
\arguments{
\item{prior.t}{list of two functions that will allow the prior density for 
stage transition times to be called}
}
\description{
Sample stage transition times from full conditional posterior
}
