% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsdive.stochit.R
\name{dsdive.stochit}
\alias{dsdive.stochit}
\title{Exploration of posterior surface via stochastic iteration}
\usage{
dsdive.stochit(depths, t, depths.labels, par, priors, max.resample, max.it,
  tol, n.paths, n.par, control.abc = NULL, control.mh = NULL,
  verbose = TRUE)
}
\arguments{
\item{depths}{Dive bins in which the trajectory was observed}

\item{t}{Times at which depths were observed}

\item{depths.labels}{character vector that defines the depth bins}

\item{par}{List with parameter values at which to initialize the MCMC 
algorithm. 
\describe{
   \item{beta}{2x3 matrix with beta parameters}
   \item{lambda}{Vector of 3 elements}
   \item{sub.tx}{vector of depth index, and a probability}
   \item{surf.tx}{Vector of two parameters}
}}

\item{max.resample}{Maximum number of times to resample imputed trajectories}

\item{max.it}{Maximum number of MCMC iterations over parameters before 
resampling imputed trajectories}

\item{tol}{Stop resampling imputed trajectories if Hellinger distance is 
below \code{tol}}

\item{n.paths}{Number of trajectories to impute}

\item{n.par}{Number of parameter values to resample when imputing 
trajectories}

\item{control.abc}{List with arguments of \code{dsdive.ldabc}, used for 
imputing complete trajectories}

\item{control.mh}{List with arguments of \code{MHadaptive::Metro_Hastings}, 
used for exploring the posterior distribution of model parameters, given 
imputed trajectories}
}
\description{
Explore the posterior distribution by updating the imputation distribution 
with samples from approximate posterior of parameters.
}
\examples{
data('dive.sim')
attach(dive.sim)

x = dsdive.stochit(depths = sim.obs$depths, t = sim.obs$times, 
                   depths.labels = depth.bins, par = params, 
                   priors = rep(5,12), max.resample = 5, max.it = 1e2, tol = .1, 
                   n.paths = 1e1, n.par = 1e1, verbose = TRUE, 
                   control.mh = list(burn_in = 1e2, quiet = FALSE))
}
