% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsdive.suffstat.R
\name{dsdive.suffstat}
\alias{dsdive.suffstat}
\title{Compute sufficient statistics for posterior distributions of complete dives}
\usage{
dsdive.suffstat(depths, durations, stages, depth.bins)
}
\description{
Compute sufficient statistics for posterior distributions of complete dives
}
