% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsdive.tx.matrix.R
\name{dsdive.tx.matrix}
\alias{dsdive.tx.matrix}
\title{Build transition matrix for dive trajectories across discrete depths}
\usage{
dsdive.tx.matrix(t0, depth.bins, beta, lambda, sub.tx, surf.tx,
  inflation.factor.lambda = 1, min.depth = 1,
  max.depth = nrow(depth.bins))
}
\arguments{
\item{t0}{time at which transition parameters should be computed}

\item{depth.bins}{\eqn{n x 2} Matrix that defines the depth bins.  The first 
column defines the depth at the center of each depth bin, and the second 
column defines the half-width of each bin.}

\item{beta}{\eqn{2 x 3} matrix in which each column contains the diving 
preference and directional persistence parameters for the DIVING, SUBMERGED, 
and SURFACING dive stages.}

\item{lambda}{length 3 vector that specifies the transition rate, 
respectively in the DIVING, SUBMERGED, and SURFACING stages.}

\item{sub.tx}{length 2 vector that specifies the first depth bin at which 
transitions to the SUBMERGED stage can occur and the probability that such 
a transition occurs at the next depth transition}

\item{surf.tx}{parameter that specifies the probability the trajectory will 
transition to the SURFACING stage at the next depth transition}
}
\description{
Dives have 3 stages, DIVING, SUBMERGED, and SURFACING.
}
\examples{
data('dive.sim')
attach(dive.sim)
attach(dive.sim$params)

m = dsdive.tx.matrix(t0 = 60, depth.bins = depth.bins, beta = beta, 
                     lambda = lambda, sub.tx = sub.tx, surf.tx = surf.tx)
}
