% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsdive.tx.params.R
\name{dsdive.tx.params}
\alias{dsdive.tx.params}
\title{Compute transition parameters for a given depth bin}
\usage{
dsdive.tx.params(depth.bins, d0, s0, beta, lambda)
}
\arguments{
\item{depth.bins}{\eqn{n x 2} Matrix that defines the depth bins.  The first 
column defines the depth at the center of each depth bin, and the second 
column defines the half-width of each bin.}

\item{d0}{index of depth bin for which to compute transition parameters}

\item{s0}{stage for which to compute parameters}

\item{beta}{vector with directional preference parameters for descent and
ascent stages.}

\item{lambda}{vector with dive rate parameters for the descent, bottom, and 
ascent stages.}
}
\description{
Computes elements of a transition matrix when given model parameters and 
time/space locations for a dive model that has 3 stages, descent, 
bottom/foraging, and ascent.
}
\examples{
data('dive.sim')
attach(dive.sim)
attach(params)

dsdive.tx.params(depth.bins = depth.bins, d0 = 3, s0 = 1, 
                 beta = beta, lambda = lambda)

detach(params)
detach(dive.sim)
}
