% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsdive.tx.stage.R
\name{dsdive.tx.stage}
\alias{dsdive.tx.stage}
\title{Compute stage transition parameters for dive trajectories across discrete depths}
\usage{
dsdive.tx.stage(t0, d0, sub.tx, surf.tx, t0.dive, t.stage2, t.scale = 60,
  rates)
}
\arguments{
\item{t0}{vector of times at which transition probs. should be computed}

\item{d0}{vector of depth bins at which transition probs. should be computed}

\item{sub.tx}{length 2 vector that specifies the first depth bin at which 
transitions to the IB stage can occur and the probability that such 
a transition occurs at the next depth transition}

\item{surf.tx}{parameter that specifies the probability the trajectory will 
transition to the PA stage at the next depth transition}

\item{t0.dive}{Time at which dive started}

\item{t.stage2}{time at which second stage was entered.  If stage 2 has not 
yet been entered, then set \code{t.stage2 = NA}, and 
\code{dsdive.tx.stage} will implicitly set \code{t.stage2 = t0}.}

\item{t.scale}{Nominally, \code{t0} and \code{t0.dive} should represent times 
in seconds relative to a starting epoch (e.g., 0 for simulations, or 
1/1/1970 for UTC date), however it may make sense to define the stage 
transition model parameters with respect to a different time scale 
(e.g., minutes, hours).  The argument \code{t.scale} will rescale \code{t0}
and \code{t0.dive} to accomplish this.}
}
\description{
Computes the probability that a dive will transition between any of the 3 
stages, PRIMARY DESCENT (PD), INTERMEDIATE BEHAVIORS (IB), and 
PRIMARY ASCENT (PA).
}
\examples{
data('dive.sim')
attach(dive.sim)
attach(params)

p = dsdive.tx.stage(t0 = sim$times, d0 = sim$depths, sub.tx = sub.tx, 
                    surf.tx = surf.tx, t0.dive = 0, t.stage2 = NA, 
                    rates = params$lambda[1], t.scale = 60)

detach(params)
detach(dive.sim)
}
