% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsdive_cdtlimpute.dsCompleteDistributed.R
\name{dsdive_cdtlimpute.dsCompleteDistributed}
\alias{dsdive_cdtlimpute.dsCompleteDistributed}
\title{Conditionally impute trajectories}
\usage{
\method{dsdive_cdtlimpute}{dsCompleteDistributed}(cfg, params, i)
}
\description{
Conditionally impute trajectories
}
