% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsdive_cdtlimpute.dsCompleteSingle.R
\name{dsdive_cdtlimpute.dsCompleteSingle}
\alias{dsdive_cdtlimpute.dsCompleteSingle}
\title{Conditionally impute trajectories}
\usage{
\method{dsdive_cdtlimpute}{dsCompleteSingle}(cfg, params, i)
}
\description{
Conditionally impute trajectories
}
