% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsdive_cdtlimpute.dsImputedBatchDistributed.R
\name{dsdive_cdtlimpute.dsImputedBatchDistributed}
\alias{dsdive_cdtlimpute.dsImputedBatchDistributed}
\title{Conditionally impute trajectories}
\usage{
\method{dsdive_cdtlimpute}{dsImputedBatchDistributed}(cfg, params, i)
}
\description{
Conditionally impute trajectories
}
