% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsdive_cdtlimpute.dsImputedDistributed.R
\name{dsdive_cdtlimpute.dsImputedDistributed}
\alias{dsdive_cdtlimpute.dsImputedDistributed}
\title{Conditionally impute trajectories}
\usage{
\method{dsdive_cdtlimpute}{dsImputedDistributed}(cfg, params, i)
}
\description{
Conditionally impute trajectories
}
