% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsdive_cdtlimpute.dsImputedLocal.R
\name{dsdive_cdtlimpute.dsImputedLocal}
\alias{dsdive_cdtlimpute.dsImputedLocal}
\title{Conditionally impute trajectories}
\usage{
\method{dsdive_cdtlimpute}{dsImputedLocal}(cfg, params, i)
}
\description{
Conditionally impute trajectories
}
