% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsdive_cdtlimpute.dsImputedSingle.R
\name{dsdive_cdtlimpute.dsImputedSingle}
\alias{dsdive_cdtlimpute.dsImputedSingle}
\title{Conditionally impute trajectories}
\usage{
\method{dsdive_cdtlimpute}{dsImputedSingle}(cfg, params, i)
}
\description{
Conditionally impute trajectories
}
