% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsdive_cdtlimpute.dssingle.R
\name{dsdive_cdtlimpute.dssingle}
\alias{dsdive_cdtlimpute.dssingle}
\title{Conditionally impute trajectories}
\usage{
\method{dsdive_cdtlimpute}{dssingle}(cfg, params, i)
}
\description{
Conditionally impute trajectories
}
