% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsdive_ld.dsImputedSingle.R
\name{dsdive_ld.dsImputedSingle}
\alias{dsdive_ld.dsImputedSingle}
\title{Evaluate the log-density of a single dive trajectory}
\usage{
\method{dsdive_ld}{dsImputedSingle}(cfg, params)
}
\arguments{
\item{cfg}{\code{dssingle} object containing dive trajectory data}

\item{params}{parameter values at which to evaluate the model's log-density}
}
\description{
Evaluates the log-density of a single dive trajectory when the computing 
environment is configured for use on a single core.
}
