% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsdive_ld.R
\name{dsdive_ld}
\alias{dsdive_ld}
\title{Define generic method for evaluating dive model likelihoods under different
computing environments}
\usage{
dsdive_ld(cfg, params)
}
\description{
All log-densities are evaluted with respect to the parameters on their 
natural scales, vs. log-densities with respect to the transformed scales.
}
