% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsdive_outputImputed.dsCompleteSingle.R
\name{dsdive_outputImputed.dsCompleteSingle}
\alias{dsdive_outputImputed.dsCompleteSingle}
\title{Define generic method for backing up gibbs sampler state under different
computing environments}
\usage{
\method{dsdive_outputImputed}{dsCompleteSingle}(cfg, output, save.time, file)
}
\description{
Define generic method for backing up gibbs sampler state under different
computing environments
}
