% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/dsdive_outputImputed.dsImputedBatchDistributed.R
\name{dsdive_outputImputed.dsImputedBatchDistributed}
\alias{dsdive_outputImputed.dsImputedBatchDistributed}
\title{Define generic method for backing up gibbs sampler state under different
computing environments}
\usage{
\method{dsdive_outputImputed}{dsImputedBatchDistributed}(cfg, output, save.time, file)
}
\description{
Define generic method for backing up gibbs sampler state under different
computing environments
}
