% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsdive_outputImputed.dsImputedDistributed.R
\name{dsdive_outputImputed.dsImputedDistributed}
\alias{dsdive_outputImputed.dsImputedDistributed}
\title{Define generic method for backing up gibbs sampler state under different
computing environments}
\usage{
\method{dsdive_outputImputed}{dsImputedDistributed}(cfg, output, save.time, file)
}
\description{
Define generic method for backing up gibbs sampler state under different
computing environments
}
