% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsdive_outputImputed.dsImputedSingle.R
\name{dsdive_outputImputed.dsImputedSingle}
\alias{dsdive_outputImputed.dsImputedSingle}
\title{Define generic method for backing up gibbs sampler state under different
computing environments}
\usage{
\method{dsdive_outputImputed}{dsImputedSingle}(cfg, output, save.time,
  file)
}
\description{
Define generic method for backing up gibbs sampler state under different
computing environments
}
