% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsdive_outputImputed.R
\name{dsdive_outputImputed}
\alias{dsdive_outputImputed}
\title{Define generic method for backing up gibbs sampler state under different
computing environments}
\usage{
dsdive_outputImputed(cfg, output, save.time, file)
}
\description{
Define generic method for backing up gibbs sampler state under different
computing environments
}
