% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envelope.approx.R
\name{envelope.approx}
\alias{envelope.approx}
\title{Approximate slope and curvature for each polynomial segment of an envelope}
\usage{
envelope.approx(breaks, anchors, lp.breaks, lp.anchors)
}
\arguments{
\item{breaks}{Breakpoints that define the start and end points of each 
interval over which a piecewise-quadratic function will be defined.}

\item{anchors}{Locations within the intervals (defined by \code{breaks}) at 
which the polynomial approximation will be centered}

\item{lp.breaks}{Value of log(f) at the breakpoints.}

\item{lp.anchors}{Value of log(f) at the anchors.}
}
\description{
Approximate slope and curvature for each polynomial segment of an envelope
}
