% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envelope.logquad.R
\name{envelope.logquad}
\alias{envelope.logquad}
\title{Build a piecewise-quadratic envelope given log of a function}
\usage{
envelope.logquad(
  breaks,
  logf,
  d.logf,
  dd.logf.sup,
  anchors = breaks[1:(length(breaks) - 1)]
)
}
\arguments{
\item{breaks}{Breakpoints that define the start and end points of each 
interval over which a piecewise-quadratic function will be defined.}

\item{logf}{Value of log(f) at the anchors.}

\item{d.logf}{Derivative of log(f) evaluated at the anchors.}

\item{dd.logf.sup}{Supremum of second derivative of log(f) over each interval}

\item{anchors}{Locations within the intervals (defined by \code{breaks}) at 
which the polynomial approximation will be centered}
}
\description{
Build a piecewise-quadratic envelope given log of a function
}
