% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffbs.segment.R
\name{ffbs.segment}
\alias{ffbs.segment}
\title{Use bridged sampling to impute a complete dive trajectory consistent with observations}
\usage{
ffbs.segment(B, x0, xN, N)
}
\arguments{
\item{B}{single-step transition matrix, after uniformization}

\item{x0}{state at which chain begins}

\item{xN}{state at which chain ends}

\item{N}{number of steps to take}
}
\description{
The sampling method is designed to sample many trajectories simultaneously, 
so has an extra level of approximation in the proposal distributions.
}
\examples{
data('dive.sim')
attach(dive.sim)
attach(dive.sim$params)

r.unif = max(outer(lambda, 2 * depth.bins$halfwidth, '/'))

m = dsdive.tx.matrix.uniformized(depth.bins = depth.bins, beta = beta, 
                                 lambda = lambda, s0 = 2, 
                                 rate.uniformized = r.unif)

N = 10

B = lapply(1:N, function(i) m)

x = ffbs.segment(B = B, x0 = 5, xN = 10, N = N)

detach(dive.sim$params)
detach(dive.sim)
}
