% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.fixedimputation.R
\name{fit.fixedimputation}
\alias{fit.fixedimputation}
\title{Gibbs sampler to estimate 3 stage dive model}
\usage{
fit.fixedimputation(par, priors = rep(5, 12), imputed.list, it,
  burn = round(it/2), verbose = FALSE)
}
\arguments{
\item{par}{List with parameter values at which to initialize the MCMC 
algorithm. 
\describe{
   \item{beta}{2x3 matrix with beta parameters}
   \item{lambda}{Vector of 3 elements}
   \item{sub.tx}{vector of depth index, and a probability}
   \item{surf.tx}{Vector of two parameters}
}}

\item{priors}{Standard deviations for mean-0 normal priors on model 
parameters, after transformation to real line}

\item{imputed.list}{A \code{list} of \code{dsdive} objects used as surrogate
data for posterior approximation.}

\item{it}{number of iterations to run MCMC sampler for}

\item{burn}{number of iterations to discard at end of sampling}

\item{verbose}{If \code{verbose=TRUE}, then print diagnostic output while 
sampling.}
}
\description{
Approximate the posterior distribution of parameters for the 3 stage dive 
model via Gibbs sampling over model parameters and imputed trajectories.
The dive bins and durations for the imputed trajectories are treated as 
fixed, but their stages are treated as latent, unknown variables that are 
sampled at each iteration if the stages are not provided.  
The full conditional posterior is used to update
the latent stages.  However, the model parameters are updated via adaptive
Metropolis-Hastings steps with a multivariate normal proposal distribution.
}
\details{
By default, the latent dive stages will be initialized by giving each dive 
stage equal time in the imputed trajectories.
}
\examples{
data('dive.sim')
attach(dive.sim)

out = fit.fixedimputation(par = params, imputed.list = list(imputed), 
                          it = 5, burn = 1, verbose = TRUE)
}
