% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexing.R
\name{fromInd}
\alias{fromInd}
\title{Converts array indices to (x,y,z) coordinates}
\usage{
fromInd(ind, x.max, y.max)
}
\arguments{
\item{ind}{index in an array}

\item{x.max}{maximum value of x coordinates}

\item{y.max}{maximum value of y coordinates}
}
\description{
Converts array indices to (x,y,z) coordinates
}
\examples{
fromInd(ind = 100, x.max = 100, y.max = 100)

}
