% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussapprox.R
\name{gaussapprox}
\alias{gaussapprox}
\title{Full conditional distribution for dive stages given data and parameters}
\usage{
gaussapprox(
  logf,
  init,
  method = "BFGS",
  optim.output = FALSE,
  lower = -Inf,
  upper = Inf,
  control = list(fnscale = -1)
)
}
\arguments{
\item{breaks}{the two indices where a new dive stage is entered}

\item{fixed.ind}{which of the stage break points in \code{breaks} should be 
held fixed}

\item{beta}{\eqn{2 x 3} matrix in which each column contains the diving 
preference and directional persistence parameters for the DIVING, SUBMERGED, 
and SURFACING dive stages.}

\item{lambda}{length 3 vector that specifies the transition rate, 
respectively in the DIVING, SUBMERGED, and SURFACING stages.}

\item{sub.tx}{length 2 vector that specifies the first depth bin at which 
transitions to the SUBMERGED stage can occur and the probability that such 
a transition occurs at the next depth transition}

\item{surf.tx}{parameter that specifies the probability the trajectory will 
transition to the SURFACING stage at the next depth transition}

\item{depths}{record of depth bins the trajectory should visit}

\item{durations}{record of amount of time spent in each depth bin}

\item{times}{times at which the depth bins should be visited}

\item{depth.bins}{\eqn{n x 2} Matrix that defines the depth bins.  The first 
column defines the depth at the center of each depth bin, and the second 
column defines the half-width of each bin.}

\item{t0.dive}{Time at which dive started}

\item{t.stage2}{time at which second stage was entered}

\item{model}{Either \code{"conditional"} or \code{"logit"} depending on the 
method used to determine stage transition probability curves}
}
\description{
The 3 stage dive model has two breakpoints where the dive stage changes.  
Given data and model parameters, this function computes the conditional 
density of one of the breakpoints given the other.  The function is useful 
for constructing a Gibbs sampler that can sample over unknown stages of a 
dive when the dive depth bins and durations are known.  For example, this 
is helpful when using the \code{crawl}-based computational strategy to 
explore the posterior distribution for model parameters.
}
\details{
Furthermore, this full conditional distribution serves as the full 
conditional posterior distribution for stage breakpoints because the stage 
breakspoints are conditionally independent from the model parameters given 
the dive depth bins and durations.
}
\examples{
library(mvtnorm)

#
# univariate example
#

# specify distribution
mu = 1
sd = 3

# get gaussian approx. to... gaussian density
g = gaussapprox(logf = function(x) dnorm(x = x, mean = mu, sd = sd, log = TRUE), 
                init = 1, method = 'Brent', lower = -10, upper = 10, 
                optim.output = TRUE)

# sample from distribution
x = g$rgaussapprox(n = 1e4)

# see that mean and sd are recovered
c(mu = mean(x), sd = sd(x))

# density matches too
max(abs(g$dgaussapprox(x = x, log = FALSE) - 
          dnorm(x = x, mean = mu, sd = sd, log = FALSE)))


#
# multivariate example
#

# set dimension
p = 3

# generate random mean
mu = rnorm(p)

# generate random covariance matrix
sigma = cov(matrix(rnorm(5*p), ncol = p))

# get gaussian approx. to... gaussian density
g = gaussapprox(
  logf = function(x) dmvnorm(x = x, mean = mu, sigma = sigma, log = TRUE), 
  init = mu + 1, method = 'BFGS', optim.output = TRUE)

# sample from distribution
x = g$rgaussapprox(n = 1e5)

# see that mean and sd are recovered
max(abs(rowMeans(x) - mu))
max(abs((cov(t(x)) - sigma)))

# density matches too
max(abs(g$dgaussapprox(x = x, log = FALSE) - 
        dmvnorm(x = t(x), mean = mu, sigma = sigma, log = FALSE)))
}
