% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs.mhga.dsdive.R
\name{gibbs.mhga.dsdive}
\alias{gibbs.mhga.dsdive}
\title{Metropolis-Hastings update for subsets of dive model parameters}
\usage{
gibbs.mhga.dsdive(
  x0,
  ld0,
  lp0,
  cfg,
  priors,
  mu.vec,
  sigma.chol,
  verbose = FALSE
)
}
\description{
The Metropolis-Hastings update in \code{gibbs.mhga.dsdive} uses the
Gaussian approximation to the posterior mode, conditional on the family of 
complete or imputed dive trajectories, as the proposal distribution.
}
