% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs.mhgdir.dsdive.R
\name{gibbs.mhgdir.dsdive}
\alias{gibbs.mhgdir.dsdive}
\title{Metropolis-Hastings update for subsets of dive model parameters}
\usage{
gibbs.mhgdir.dsdive(x0, ld0, lp0, cfg, priors, verbose = FALSE)
}
\description{
The Metropolis-Hastings update in \code{gibbs.mhga.dsdive} uses the
Gaussian approximation to the posterior mode, conditional on 1) the family of 
complete or imputed dive trajectories, and 2) a search direction in which to 
move, as the proposal distribution.
}
