% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeCompleteSingle.R
\name{makeCompleteSingle}
\alias{makeCompleteSingle}
\title{Initialize a computing environment for use with gibbs sampling}
\usage{
makeCompleteSingle(depth.bins, durations, depths, times, stages, init,
  priors, t0.dive)
}
\description{
Initialize a computing environment for use with gibbs sampling
}
\examples{
data('dive.sim')
attach(dive.sim)
attach(dive.sim$params)

spec = list(beta.sd = rep(1, 3), beta.absmax = 5, lambda.sd = 1, 
            sub.tx.mean = 0, sub.tx.sd = 1, surf.tx.mean = 0, surf.tx.sd = 1)

cfg = makeCompleteSingle(depth.bins = depth.bins, durations = sim$durations, 
                         depths = sim$depths, times = sim$times, 
                         stages = sim$stages, init = params, priors = spec, 
                         t0.dive = 0)

detach(dive.sim$params)
detach(dive.sim)
}
