% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeImputedSingle.R
\name{makeImputedSingle}
\alias{makeImputedSingle}
\title{Initialize a computing environment for use with gibbs sampling}
\usage{
makeImputedSingle(depth.bins, it, depths, times, init, priors,
  inflation.factor.lambda, t0.dive, verbose = FALSE)
}
\description{
Initialize a computing environment for use with gibbs sampling
}
\examples{
data('dive.sim')
attach(dive.sim)
attach(dive.sim$params)

spec = list(beta.sd = rep(1, 3), beta.absmax = 5, lambda.sd = 1, 
            sub.tx.mean = 0, sub.tx.sd = 1, surf.tx.mean = 0, surf.tx.sd = 1)

cfg = makeImputedSingle(depth.bins = depth.bins, it = 5, 
                        depths = sim.obs$depths, times = sim.obs$times, 
                        init = params, priors = spec, 
                        inflation.factor.lambda = 1.1, t0.dive = 0, 
                        verbose = TRUE)

x = dsdive.fit.gibbs.cfg(cfg = cfg, it = 2, verbose = TRUE, init = params, 
                         priors = spec, sigma = list(diag(3), diag(7)))

detach(dive.sim$params)
detach(dive.sim)
}
