% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSingle.R
\name{makeSingle}
\alias{makeSingle}
\title{Initialize a computing environment for use with gibbs sampling}
\usage{
makeSingle(depth.bins, durations, it, depths, times, init, priors,
  inflation.factor.lambda, t0.dive, verbose = FALSE)
}
\description{
Initialize a computing environment for use with gibbs sampling
}
