% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhrw.R
\name{mhrw}
\alias{mhrw}
\title{Metropolis-Hastings random walk proposal and update}
\usage{
mhrw(x0, lp, cov.chol)
}
\arguments{
\item{x0}{Initial location of sampler}

\item{lp}{Function to evaluate log-posterior at model parameters}

\item{cov.chol}{Cholesky decomposition of proposal covariance}
}
\description{
Sampler uses a mean-zero Gaussian random walk as a proposal distribution.
}
