% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.convert.R
\name{params.toVec}
\alias{params.toVec}
\title{Convert list of parameters to vector of parameters on transformed}
\usage{
params.toVec(par, spec)
}
\arguments{
\item{par}{the list of parameters to convert to a vector of model parameters}

\item{spec}{List of parameters to specify prior distributions.  See 
\code{dsdive.prior} for more details.}
}
\description{
Convert list of parameters to vector of parameters on transformed
}
\examples{
data('dive.sim')
attach(dive.sim)
attach(params)

spec = list(beta.sd = 1, beta.absmax = 10, lambda.sd = 1, sub.tx.mean = 0, 
            sub.tx.sd = 1, surf.tx.mean = 0, surf.tx.sd = 1)

v = params.toVec(par = params, spec = spec)

L = params.toList(par = v, spec = spec)

detach(params)
detach(dive.sim)
}
