% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.dsdive.R
\name{plot.dsdive}
\alias{plot.dsdive}
\title{Plotting functions for completely observed dive trajectories}
\usage{
\method{plot}{dsdive}(x, depth.bins, dsobs = NULL, imputed.list = NULL,
  imputed.alpha = 0.3, ...)
}
\arguments{
\item{x}{\code{dsdive} object with complete record of dive trajectory}

\item{depth.bins}{\eqn{n x 2} Matrix that defines the depth bins.  The first 
column defines the depth at the center of each depth bin, and the second 
column defines the half-width of each bin.}

\item{dsobs}{\code{dsobs} object with record of observed depth bins}

\item{imputed.list}{}
}
\description{
Plotting functions for completely observed dive trajectories
}
\examples{
data('dive.sim')
attach(dive.sim)

# basic plot
pl = plot(x = sim.obs, depth.bins = depth.bins)

# plot with stages overlaid
pl2 = plot(x = sim.obs, depth.bins = depth.bins, stages = sim.obs$stages)
}
