% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.dsobs.R
\name{plot.dsobs}
\alias{plot.dsobs}
\title{Plotting functions for observations of dive trajectories}
\usage{
\method{plot}{dsobs}(x, depth.bins, stages = NULL, errorbars = FALSE,
  ...)
}
\arguments{
\item{x}{\code{dsobs} object with record of observed depth bins}

\item{depth.bins}{Vector that defines the depth bins (i.e., bin labels)}

\item{stages}{Optional vector of for which dive stage the trajectory was 
in at each observation}

\item{errorbars}{If \code{TRUE}, then the minimum and maximum depth for each
bin will be included in the plot.}
}
\description{
Plotting functions for observations of dive trajectories
}
\examples{
data('dive.sim')
attach(dive.sim)

# basic plot
pl = plot(x = sim.obs, depth.bins = depth.bins)

# plot with stages overlaid
stages = c(rep(1, 5), rep(2,15), rep(3,6))
pl2 = plot(x = sim.obs, depth.bins = depth.bins, stages = stages)
}
