% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proposal.logit.R
\name{proposal.logit}
\alias{proposal.logit}
\title{Logit-based Metropolis-Hastings Random-Walk proposals}
\usage{
proposal.logit(x0, sd, a = 0, b = 1)
}
\arguments{
\item{x0}{initial parameter value}

\item{sd}{standard deviation of proposal, on logit scale}

\item{a}{minimum allowable value for the proposal}
}
\description{
Uses a logit transform on x0 to propose a new parameter value, for use in 
MH RW Gibbs steps.  The function also returns the log-ratio of the proposal.
}
\examples{
proposal.logit(x0 = 2, sd = .1, a = 0, b = 10)
}
