% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refine.partition.R
\name{refine.partition}
\alias{refine.partition}
\title{Refine a partition of an interval}
\usage{
refine.partition(breaks, max.width)
}
\arguments{
\item{breaks}{Initial partition of an interval on the real line}

\item{max.width}{Maximum width of intervals in output partition

# @example examples/refine.partition.R}
}
\description{
The input \code{breaks} defines a partition.  \code{refine.partition} will 
add evenly spaced points between \code{breaks} such that the maximum 
difference between any two points in the output is \code{max.width}.
}
