% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rejection.sample.R
\name{rejection.sample}
\alias{rejection.sample}
\title{Rejection sampler}
\usage{
rejection.sample(n, log.f, log.g, rg, a, max.it = 1000)
}
\arguments{
\item{n}{number of samples to draw}

\item{log.f}{log of target density}

\item{log.g}{log of envelope kernel}

\item{rg}{function to draw samples from g; should only accept "n" as an 
argument}

\item{a}{scaling factor for envelope.  a is such that g/a > f everywhere.}

\item{max.it}{maximum number of tries before failing to generate sample}
}
\value{
a random draw from f, or NA if sampling fails
}
\description{
Rejection sampler
}
\examples{
# number of samples to draw
n = 1e4

unif.circle = rejection.sample(
  n = n, 
  log.f = function(x) { log(sum(x^2) < 1) + log(pi) }, 
  log.g = function(x) 0, 
  rg = function(n) runif(2, min = -1, max = 1), 
  a = pi/4)

apply(unif.circle, 1, mean)

plot(t(unif.circle), xlim = c(-1,1), ylim = c(-1,1), asp = 1, 
     xlab = expression(X[1]), ylab = expression(X[2]))

}
