% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spsa.R
\name{spsa}
\alias{spsa}
\title{Basic implementation of stochastic optimization when the target function is only approximately known}
\usage{
spsa(par, fn, maxit = 100, a = 1, c = 1, A = 1, alpha = 0.5,
  gamma = 1, tol = 0.001)
}
\arguments{
\item{par}{initial parameter values for function}

\item{fn}{a function to be maximized}

\item{maxit}{maximum number of function evaluations before failing to converge}

\item{a}{decay parameter for step size}

\item{c}{decay parameter for step size}

\item{A}{decay parameter for step size}

\item{alpha}{decay parameter for step size}

\item{gamma}{decay parameter for step size}

\item{tol}{Relative convergence criterion

# @example examples/spsa.R}
}
\description{
Experimental method because its theoretical convergence and recommendations 
for setting parameters has not been fully explored.
}
