% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stagevec.R
\name{stagevec}
\alias{stagevec}
\title{Convenience function for building a vector for dive stages}
\usage{
stagevec(length.out, breaks)
}
\arguments{
\item{length.out}{Length of vector to generate}

\item{breaks}{Location of changepoints}
}
\description{
Several functions in the \code{dsdive} package have a record of dive 
stages as one of the function's arguments.  However, this information is not 
always available.  For example, dive stage information is a latent variable
so is not included with observational data.
}
\details{
The \code{stagevec} function makes it easy to generate a vector of dive 
stages, like \code{c(1,1,1,2,2,2,2,3,3)}.  Users only need to provide a 
target length \code{length.out} and the indices at which the stage 
changes \code{breaks.}
}
\examples{

stagevec(length.out = 30, breaks = c(7, 20))

}
