% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theta.grid.R
\name{theta.grid}
\alias{theta.grid}
\title{Build parameter grid in the proper enumeration order}
\usage{
theta.grid(pi1, pi2, lambda1, lambda2, lambda3)
}
\description{
Build parameter grid in the proper enumeration order
}
