% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexing.R
\name{toInd}
\alias{toInd}
\title{Converts (x,y,z) coordinates to an array index}
\usage{
toInd(x, y, z, x.max, y.max)
}
\arguments{
\item{x}{x coordinate, ranges between 1 and \code{x.max}}

\item{y}{y coordinate, ranges between 1 and \code{y.max}}

\item{z}{z coordinate, must be at least 1}

\item{x.max}{maximum value of x coordinates}

\item{y.max}{maximum value of y coordinates}
}
\description{
Used for getting the flattened index for an array with dimensions
\eqn{x.max x y.max x n}, where \eqn{n} is arbitrary.
}
\examples{
toInd(x = 60, y = 13, z = 1, x.max = 100, y.max = 100)

}
