% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dive.sim}
\alias{dive.sim}
\title{Simulated three stage dive}
\format{An object of class \code{list} of length 5.}
\usage{
dive.sim
}
\description{
The \code{dive.sim} object contains the complete simulated dive trajectory 
and as a corresponding set of observations in addition to the model 
parameters and other features used to generate and analyze the dive.  The 
purpose of the dataset is to provide illustrations of data formats 
compatible with the \code{dsdive} package, and also provide pre-computed 
outputs that may be worked with immediately when exploring and testing the 
package.
}
\examples{

data("dive.sim")
str(dive.sim)

}
\keyword{datasets}
