% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.df.R
\name{ds.df}
\alias{ds.df}
\title{Builds a tidy data frame for dive objects}
\usage{
ds.df(depths, times, depth.bins, stages = NULL, durations = NULL,
  time.as.POSIXct = FALSE, t0.dive = times[1])
}
\arguments{
\item{depths}{Record of observed depth bin indices}

\item{times}{Times (in seconds) at which depth bin observations were made}

\item{depth.bins}{\eqn{n x 2} Matrix that defines the depth bins.  The first 
column defines the depth at the center of each depth bin, and the second 
column defines the half-width of each bin.}

\item{stages}{Vector of (guesses) for which dive stage the trajectory was 
in at each observation}

\item{durations}{Vector specifying the amount of time spent in each depth bin}

\item{time.as.POSIXct}{if \code{TRUE}, will convert plotting times to POSIXct}

\item{t0.dive}{Initial time for dive, so that times from beginning of dive 
can be computed as well}
}
\description{
Converts raw dive information to a tidy dataframe that
is suitable for plotting with \code{ggplot2}.  Most importantly, the function
converts depth bin indices to actual depths and depth ranges.
}
\examples{
data('dive.sim')
attach(dive.sim)

df = ds.df(depths = sim.obs$depths, times = sim.obs$times, 
           depth.bins = depth.bins, stages = NULL)

detach(dive.sim)
}
