% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsdive.align.obs.R
\name{dsdive.align.obs}
\alias{dsdive.align.obs}
\title{Extract a segment of a fully observed dive trajectory between timepoints}
\usage{
dsdive.align.obs(depths, times, t.stages, offset, offset.tf)
}
\arguments{
\item{depths}{Complete record of depth bin indices visited}

\item{times}{Times at which each of \code{depths} was visited}

\item{stages}{Stages at which each of the \code{depths} was visited}

\item{durations}{the amount of time spent in each depth bin}

\item{t0}{the time to mark the beginning of the segment to extract}

\item{tf}{the time to mark the end of the segment to extract}
}
\description{
Extract a segment of a fully observed dive trajectory between timepoints
}
\examples{
data('dive.sim')
attach(dive.sim)
attach(params)

t.stages = sim$times[c(FALSE, diff(sim$stages)==1)]
  
dsdive.align.obs(depths = sim.obs$depths, times = sim.obs$times, 
                 t.stages = t.stages, offset = -300, 
                 offset.tf = -300)

detach(params)
detach(dive.sim)
}
