% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsdive.fwdsample.dive.R
\name{dsdive.fwdsample.dive}
\alias{dsdive.fwdsample.dive}
\title{Simulate dive trajectories across discrete depth bins}
\usage{
dsdive.fwdsample.dive(depth.bins, beta, lambda, t0, steps.max, T1, T2)
}
\arguments{
\item{depth.bins}{\eqn{n x 2} Matrix that defines the depth bins.  The first 
column defines the depth at the center of each depth bin, and the second 
column defines the half-width of each bin.}

\item{beta}{Directional preference model parameters.  See 
\code{dsdive.tx.params} for more details.}

\item{lambda}{Diving rate model parameters.  See 
\code{dsdive.tx.params} for more details.}

\item{t0}{Initial time for the dive}

\item{steps.max}{maximum number of transitions to sample before stopping, 
regardless of whether dive has finished.}

\item{T1}{stage 1 duration}

\item{T2}{stage 2 duration}
}
\value{
A \code{dsdive} object, which is a \code{list} with the following 
  vectors:
  \describe{
    \item{depths}{Record of which depth bin indices the trajectory visited}
    \item{durations}{Record of amount of time spent in each depth bin}
    \item{times}{The time at which each depth bin was entered}
    \item{stages}{The stage at which each depth bin was entered}
  }
}
\description{
The method will simulate a complete dive from its beginning to end.
}
\examples{
data('dive.sim')
attach(dive.sim)
attach(dive.sim$params)

# simulate dive
x = dsdive.fwdsample.dive(depth.bins = depth.bins, beta = beta, 
                          lambda = lambda, t0 = 0, steps.max = 1e3, 
                          T1 = 15*60, T2 = 12*60)

pl = plot(x = x, depth.bins = depth.bins)

detach(dive.sim$params)
detach(dive.sim)
}
