% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsdive.fwdsample.R
\name{dsdive.fwdsample}
\alias{dsdive.fwdsample}
\title{Simulate dive trajectories across discrete depth bins}
\usage{
dsdive.fwdsample(depth.bins, d0, beta, lambda, sub.tx, surf.tx, t0, tf,
  steps.max, dur0 = NULL, nsteps = NULL, s0, t0.dive,
  shift.params = NULL, t.stage2)
}
\arguments{
\item{depth.bins}{\eqn{n x 2} Matrix that defines the depth bins.  The first 
column defines the depth at the center of each depth bin, and the second 
column defines the half-width of each bin.}

\item{d0}{the depth bin at which transition parameters should be computed}

\item{beta}{\eqn{2 x 3} matrix in which each column contains the diving 
preference and directional persistence parameters for the DIVING, SUBMERGED, 
and SURFACING dive stages.}

\item{lambda}{length 3 vector that specifies the transition rate, 
respectively in the DIVING, SUBMERGED, and SURFACING stages.}

\item{sub.tx}{length 2 vector that specifies the first depth bin at which 
transitions to the SUBMERGED stage can occur and the probability that such 
a transition occurs at the next depth transition}

\item{surf.tx}{parameter that specifies the probability the trajectory will 
transition to the SURFACING stage at the next depth transition}

\item{t0}{time at which transition parameters should be computed}

\item{tf}{time at which sampling should end after}

\item{steps.max}{maximum number of transitions to sample before stopping, 
regardless of whether \code{tf} is reached.}

\item{dur0}{time spent at location \code{d0}.  If \code{NULL}, then a 
duration in state \code{d0} will be sampled, otherwise a new state will 
be sampled first, then sampling will continue from the new state at time 
\code{t0 + dur0}.}

\item{nsteps}{if \code{nsteps} is not \code{NULL}, then the sampler will
be reconfigured to sample exactly \code{nsteps} transitions instead of 
attempting to sample until the trajectory is observable at time \code{tf}.}

\item{s0}{dive stage in which forward simulation begins}

\item{t0.dive}{Time at which dive started}

\item{shift.params}{Optional arguments to bias sampling toward a specific 
node.  See documentation for \code{dsdive.tx.params} for more detail.}
}
\value{
A \code{dsdive} object, which is a \code{list} with the following 
  vectors:
  \describe{
    \item{depths}{Record of which depth bin indices the trajectory visited}
    \item{durations}{Record of amount of time spent in each depth bin}
    \item{times}{The time at which each depth bin was entered}
    \item{stages}{The stage at which each depth bin was entered}
  }
}
\description{
The method will simulate dive trajectories from initial conditions until the 
trajectory is observable at \code{tf}, or a maximum number of transitions 
has been exceeded.  The dive simulation is bridged, so the trajectory will
also stop diving after returning to the surface.
}
\examples{
data('dive.sim')
attach(dive.sim)
attach(dive.sim$params)

# simulate dive
x = dsdive.fwdsample(depth.bins = depth.bins, d0 = 1, beta = beta, 
                     lambda = lambda, sub.tx = sub.tx, surf.tx = surf.tx, 
                     t0 = 0, tf = Inf, steps.max = 1e5, dur0 = NULL, 
                     nsteps = NULL, s0 = 1, t0.dive = 0, t.stage2 = NA)

detach(dive.sim$params)
detach(dive.sim)
}
