% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsdive.gibbs.obs.R
\name{dsdive.gibbs.obs}
\alias{dsdive.gibbs.obs}
\title{Likelihood for completely observed dive trajectories}
\usage{
dsdive.gibbs.obs(
  dsobs.list,
  t.stages.list,
  beta.init,
  lambda.init,
  verbose = FALSE,
  maxit,
  checkpoint.fn,
  checkpoint.interval = 3600,
  pi1.prior,
  pi2.prior,
  lambda1.prior,
  lambda2.prior,
  lambda3.prior,
  tstep,
  depth.bins,
  T1.prior.params,
  T2.prior.params,
  max.width
)
}
\arguments{
\item{depth.bins}{\eqn{n x 2} Matrix that defines the depth bins.  The first 
column defines the depth at the center of each depth bin, and the second 
column defines the half-width of each bin.}

\item{depths}{Depth bin indices in which the trajectory was observed}

\item{durations}{Amount of time spent in each depth bin}

\item{times}{Times at which each depth bin was entered}

\item{stages}{Record of dive stages at each depth bin}

\item{beta}{\eqn{2 x 3} matrix in which each column contains the diving 
preference and directional persistence parameters for the DIVING, SUBMERGED, 
and SURFACING dive stages.}

\item{lambda}{length 3 vector that specifies the transition rate, 
respectively in the DIVING, SUBMERGED, and SURFACING stages.}

\item{sub.tx}{length 2 vector that specifies the first depth bin at which 
transitions to the SUBMERGED stage can occur and the probability that such 
a transition occurs at the next depth transition}

\item{surf.tx}{parameter that specifies the probability the trajectory will 
transition to the SURFACING stage at the next depth transition}

\item{t0.dive}{Time at which dive started}

\item{d0.last}{If the depth bin that proceeded the first depth bin in 
\code{depths}.  If the trajectory to be analyzed was started at the 
surface, then set \code{c0.last=NULL}.}

\item{model}{Either \code{"conditional"} or \code{"logit"} depending on the 
method used to determine stage transition probability curves}
}
\description{
Likelihood for completely observed dive trajectories
}
\examples{
data('dive.sim')
attach(dive.sim)
attach(dive.sim$params)

t.stages = sim$times[c(FALSE,diff(sim$stages)==1)]
  
tstep = diff(sim.obs$times[1:2])

obstx.mat = lapply(1:3, function(s) {
  dsdive.obstx.matrix(depth.bins = depth.bins, beta = beta, 
                      lambda = lambda, s0 = s, tstep = tstep, 
                      include.raw = TRUE)
})

lambda.priors = list(
  c(4, 2 + 2/3), c(4, 13 + 1/3), c(4, 5)
)

beta.priors = list(
  c(31.5, 3.5), c(3.5, 31.5)
)

T1.prior.params = c(25, .04)
T2.prior.params = c(56, .06)

fit = dsdive.gibbs.obs(
  dsobs.list = list(sim.obs), t.stages.list = list(t.stages), 
  beta.init = beta, lambda.init = lambda, verbose = TRUE, maxit = 1e3, 
  pi1.prior = beta.priors[[1]], pi2.prior = beta.priors[[2]], 
  lambda1.prior = lambda.priors[[1]], lambda2.prior = lambda.priors[[2]], 
  lambda3.prior = lambda.priors[[3]], tstep = tstep, depth.bins = depth.bins, 
  T1.prior.params = T1.prior.params, T2.prior.params = T2.prior.params, 
  max.width = 100)


detach(dive.sim$params)
detach(dive.sim)
}
