% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsdive.impute.R
\name{dsdive.impute}
\alias{dsdive.impute}
\title{Use bridged sampling to impute a complete dive trajectory consistent with observations}
\usage{
dsdive.impute(depth.bins, depths, times, s0, beta, lambda, sub.tx, surf.tx,
  inflation.factor.lambda = 1.1, verbose = FALSE,
  precompute.bridges = TRUE)
}
\arguments{
\item{depth.bins}{\eqn{n x 2} Matrix that defines the depth bins.  The first 
column defines the depth at the center of each depth bin, and the second 
column defines the half-width of each bin.}

\item{depths}{record of depth bins the trajectory should visit}

\item{times}{times at which the depth bins should be visited}

\item{s0}{dive stage at which the trajectory should be started from}

\item{beta}{\eqn{2 x 3} matrix in which each column contains the diving 
preference and directional persistence parameters for the DIVING, SUBMERGED, 
and SURFACING dive stages.}

\item{lambda}{length 3 vector that specifies the transition rate, 
respectively in the DIVING, SUBMERGED, and SURFACING stages.}

\item{sub.tx}{length 2 vector that specifies the first depth bin at which 
transitions to the SUBMERGED stage can occur and the probability that such 
a transition occurs at the next depth transition}

\item{surf.tx}{parameter that specifies the probability the trajectory will 
transition to the SURFACING stage at the next depth transition}

\item{inflation.factor.lambda}{In order to facilitate bridged transitions, 
the transition rate of the overall process must be inflated to allow the 
possibility of self-transitions.  Self-transitions allow bridged paths to 
dynamically modify the total number of transitions between observed values
so that a valid path between observations is always possible.  The 
\code{inflation.factor.lambda} parameter implicitly controls the number of 
self-transitions that will occur.  Larger values will create more 
self-transitions.}

\item{verbose}{If \code{TRUE}, then the sampler's progress will be printed 
during sampling.}

\item{precompute.bridges}{If \code{TRUE}, then the bridged transition 
matrices will be precomputed.  Enabling this option will increase the 
memory overhead of the method, but will reduce its runtime.}
}
\description{
Use bridged sampling to impute a complete dive trajectory consistent with observations
}
\examples{
data('dive.sim')
attach(dive.sim)
attach(params)

imputed = dsdive.impute(depth.bins = depth.bins, depths = sim.obs$depths, 
                        times = sim.obs$times, s0 = 1, beta = beta, 
                        lambda = lambda, sub.tx = sub.tx, surf.tx = surf.tx)
}
