% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsdive.ldabc.R
\name{dsdive.ldabc}
\alias{dsdive.ldabc}
\title{SMC estimation of likelihood for imputed dive trajectories}
\usage{
dsdive.ldabc(beta, lambda, sub.tx, surf.tx, depths.labels,
  steps.max = 1000, N, depths, t, tries.max, eps, dump.state = FALSE,
  verbose = FALSE, n.samples = 0, stage.init = 1)
}
\arguments{
\item{steps.max}{maximum number of transitions to sample before stopping, 
regardless of whether \code{tf} is reached.}

\item{N}{number of SMC particles to maintain}

\item{depths}{Dive bins in which the trajectory was observed}

\item{t}{vector of times at which observations were made}

\item{tries.max}{maximum number of attempts to find a suitable ABC trajectory 
before stopping sampling.  If \code{tries.max} is exceeded while sampling 
any location, the returned log-likelihood estimate will be \code{-Inf}}

\item{eps}{ABC rejection sampling tolerance; maximum distance along graph 
between simulated and observed states}

\item{dump.state}{if sampling fails and \code{dump.state==TRUE}, then the 
particles in the particle filter will be returned, along with the timepoint 
at which sampling failed}

\item{verbose}{if \code{TRUE}, then various diagnostic or informational 
output will be printed while the ABC-SMC algorithm is running}

\item{n.samples}{the number of imputed trajectories to return.  If 
\code{n.samples==0}, then no trajectories will be saved during sampling, 
which can potentially yield faster inference because each step of the 
ABC-SMC sampler requires fewer memory operations.}

\item{Xstatic}{\code{(num. nodes) x (num. local covariates)} matrix
containing location-based drivers of movement}

\item{betaStatic}{\code{(num. local covariates)} long vector containign 
coefficients associated with location-based drivers of movement}

\item{beta_ar}{a number specifying the strength of the autoregressive effect
for the CTDS model.  If there is no autoregressive component, setting 
\code{beta_ar = NULL} will skip the associated computations.}

\item{Q}{sparse matrix that defines the discrete domain and connections 
between areal units}

\item{coords}{\code{(num. nodes) x (num. dimensions)} matrix containing 
coordinates for each node; used for computing directional drivers between 
locations.}

\item{states}{vector of locations at which CTDS trajectory was observed}

\item{nb}{list that defines the discrete domain and connections 
between areal units.  each entry contains a vector of neighbors for each 
location.}

\item{t}{Times at which depths were observed}
}
\description{
SMC estimation of likelihood for imputed dive trajectories
}
\examples{
data('dive.sim')
attach(dive.sim)
attach(dive.sim$params)

ld = dsdive.ldabc(beta = beta, lambda = lambda, sub.tx = sub.tx, 
                  surf.tx = surf.tx, depths.labels = depth.bins, 
                  steps.max = 1e3, N = 1e1, depths = sim.obs$depths, 
                  t = sim.obs$times, tries.max = 1e5, dump.state = TRUE, 
                  verbose = TRUE, n.samples = 5, eps = 1)
}
