% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsdive.obs.sampleparams.R
\name{dsdive.obs.sampleparams}
\alias{dsdive.obs.sampleparams}
\title{Sample pi and lambda parameters for a single dive stage}
\usage{
dsdive.obs.sampleparams(
  dsobs.list,
  t.stages.list,
  P.raw,
  s0,
  depth.bins,
  beta,
  lambda,
  lambda.priors.list,
  beta.priors.list,
  tstep,
  gapprox = NULL,
  output.gapprox = FALSE
)
}
\arguments{
\item{dsobs.list}{list of \code{dsobs} objects, which describe the 
observation times and depths of a collection of dives}

\item{t.stages.list}{list of initial stage transition times for dives 
observed in \code{dsobs.list}}

\item{P.raw}{list of continuous time probability transition matrices, and 
components.}

\item{s0}{stage for which model parameters should be sampled}

\item{depth.bins}{\eqn{n x 2} Matrix that defines the depth bins.  The first 
column defines the depth at the center of each depth bin, and the second 
column defines the half-width of each bin.}

\item{beta}{directional preference model parameters.
See \code{dsdive.tx.params} for more details.}

\item{lambda}{diving rate model parameters.  See 
\code{dsdive.tx.params} for more details.}

\item{lambda.priors.list}{list of parameters for the prior distributions for 
the diving rate parameters.  See \code{dsdive.gibbs.obs} for more 
information.}

\item{beta.priors.list}{list of parameters for the prior distributions for 
the directional preference parameters.  See \code{dsdive.gibbs.obs} for  
more information.  Note that these priors are specified as "pi" priors vs. 
priors for "beta".}

\item{tstep}{Time between observations in \code{dsobs.unaligned}}

\item{gapprox}{(Optional) \code{gaussapprox} object containing the Gaussian 
approximation used to propose model parameters.  If \code{NULL}, then a 
Gaussian approximation will be computed.}

\item{output.gapprox}{\code{TRUE} to return the Gaussian approximation used 
to propose model parameters.}
}
\description{
Sampler uses a Gaussian approximation to the full conditional posterior.
}
\examples{
data('dive.sim')
attach(dive.sim)
attach(dive.sim$params)

t.stages = sim$times[c(FALSE,diff(sim$stages)==1)]
  
tstep = diff(sim.obs$times[1:2])

obstx.mat = lapply(1:3, function(s) {
  dsdive.obstx.matrix(depth.bins = depth.bins, beta = beta, 
                      lambda = lambda, s0 = s, tstep = tstep, 
                      include.raw = TRUE)
})

lambda.priors = list(
  c(4, 2 + 2/3), c(4, 13 + 1/3), c(4, 5)
)

beta.priors = list(
  c(31.5, 3.5), c(3.5, 31.5)
)

x = dsdive.obs.sampleparams(dsobs.list = list(sim.obs), 
                            t.stages.list = list(t.stages), P.raw = obstx.mat, 
                            s0 = 1, depth.bins = depth.bins, beta = beta, 
                            lambda = lambda, lambda.priors.list = lambda.priors, 
                            beta.priors.list = beta.priors, 
                            tstep = diff(sim.obs$times[1:2]))

detach(dive.sim$params)
detach(dive.sim)
}
