% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsdive.observe.R
\name{dsdive.observe}
\alias{dsdive.observe}
\title{Simulate observing a fully observed dive trajectory at specified timepoints}
\usage{
dsdive.observe(depths, times, stages, t.obs)
}
\arguments{
\item{depths}{Complete record of depth bin indices visited}

\item{times}{Times at which each of \code{depths} was visited}

\item{stages}{Stages at which each of the \code{depths} was visited}

\item{t.obs}{Times at which the trajectory should be observed}
}
\description{
Given the components of a completely observed dive trajectory, extract the 
depth and stage of a trajectory at the target times \code{t.obs}.
}
\examples{
data('dive.sim')
attach(dive.sim)

obs = dsdive.observe(depths = sim$depths, times = sim$times, 
                     stages = sim$stages, 
                     t.obs = seq(from = 0, to = max(sim$times)+60, by = 60))
}
