% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsdive.obstx.matrix_interpolator.R
\name{dsdive.obstx.matrix_interpolator}
\alias{dsdive.obstx.matrix_interpolator}
\title{Interpolators for probability transition matrices of partially observed CTMC}
\usage{
dsdive.obstx.matrix_interpolator(
  depth.bins,
  beta.seq,
  lambda.seq,
  s0,
  tstep.seq,
  m,
  verbose = FALSE
)
}
\arguments{
\item{depth.bins}{\eqn{n x 2} Matrix that defines the depth bins.  The first 
column defines the depth at the center of each depth bin, and the second 
column defines the half-width of each bin.}

\item{beta.seq}{the range of depth bin transition model parameters}

\item{lambda.seq}{the range of depth bin transition rate model parameters}

\item{s0}{the stage for which to compute the transition matrix}

\item{tstep.seq}{Range of times between observations of the CTMC}

\item{m}{Smoothing polynomial degree; this is the \code{m} argument in the 
smoothing function, \code{fields::Tps}.}

\item{verbose}{\code{TRUE} to output progress of code}
}
\description{
Given a range of model parameters, this function will compute the 
probability transition matrices for all combinations of a Continuous time 
Markov chain (CTMC) that is observed once, and  then again at a time of 
\code{tstep} units of time later.  Then, thin plate spline interpolating 
functions will be built to allow probability transition matrix entries to 
be approximated at different values of the input parameters.
}
\examples{
data('dive.sim')
attach(dive.sim)

# support for descent directional preferences
beta.seq = seq(.5, 1, length.out = 7)
beta.seq[length(beta.seq)] = .999

# support for descent speeds
lambda.seq = seq(.1, 2, length.out = 7)

# support for timesteps
tstep.seq = seq(0, 300, by = 100)

# build interpolating function
interpolator = dsdive.obstx.matrix_interpolator(
  depth.bins = depth.bins, beta.seq = beta.seq, lambda.seq = lambda.seq, 
  s0 = 1, tstep.seq = tstep.seq, m = 3, verbose = FALSE)

interpolator(beta = .8, lambda = 1, tstep = 300, i = 1, j = 4)

detach(dive.sim)
}
