% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsdive.sample.stages.R
\name{dsdive.sample.stages}
\alias{dsdive.sample.stages}
\title{Sample stage transition times from full conditional posterior}
\usage{
dsdive.sample.stages(
  depths,
  durations,
  times,
  t.stages,
  beta,
  lambda,
  depth.bins,
  T1.prior,
  T2.prior
)
}
\arguments{
\item{prior.t}{list of two functions that will allow the prior density for 
stage transition times to be called}
}
\description{
Sample stage transition times from full conditional posterior
}
\examples{
data('dive.sim')
attach(dive.sim)
attach(dive.sim$params)

t.stages = sim$times[diff(sim$stages) == 1]

T1.prior = function(x) dgamma(x, shape = 15, rate = 1, log = TRUE)
T2.prior = function(x) dgamma(x, shape = 15, rate = 1, log = TRUE)

dsdive.sample.stages(depths = sim$depths, durations = sim$durations, 
                     times = sim$times, t.stages = t.stages, beta = beta, 
                     lambda = lambda, depth.bins = depth.bins, 
                     T1.prior = T1.prior, T2.prior = T2.prior)

detach(dive.sim$params)
detach(dive.sim)
}
