% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsdive.stochit.R
\name{dsdive.stochit}
\alias{dsdive.stochit}
\title{Exploration of posterior surface via stochastic iteration}
\usage{
dsdive.stochit(depths, t, depth.bins, par, priors, max.resample, max.it,
  tol, n.paths, n.par, control.abc = NULL, control.mh = NULL,
  verbose = TRUE)
}
\arguments{
\item{depths}{Dive bins in which the trajectory was observed}

\item{t}{Times at which depths were observed}

\item{depth.bins}{\eqn{n x 2} Matrix that defines the depth bins.  The first 
column defines the depth at the center of each depth bin, and the second 
column defines the half-width of each bin.}

\item{par}{List with parameter values at which to initialize the MCMC 
algorithm. 
\describe{
   \item{beta}{2x3 matrix with beta parameters}
   \item{lambda}{Vector of 3 elements}
   \item{sub.tx}{vector of depth index, and a probability}
   \item{surf.tx}{Vector of two parameters}
}}

\item{priors}{Standard deviations for mean-0 normal priors on model 
parameters, after transformation to real line}

\item{max.resample}{Maximum number of times to resample imputed trajectories}

\item{max.it}{Maximum number of MCMC iterations over parameters before 
resampling imputed trajectories}

\item{tol}{Stop resampling imputed trajectories if Hellinger distance is 
below \code{tol}}

\item{n.paths}{Number of trajectories to impute}

\item{n.par}{Number of parameter values to resample when imputing 
trajectories}

\item{control.abc}{List with arguments of \code{dsdive.ldabc}, used for 
imputing complete trajectories}

\item{control.mh}{List with arguments of \code{MHadaptive::Metro_Hastings}, 
used for exploring the posterior distribution of model parameters, given 
imputed trajectories}
}
\description{
Experimental method to explore the posterior distribution by approximating 
the posterior distribution by running MCMC over a family of imputed 
trajectories.  After the MCMC chain completes, the family of imputed 
trajectories is updated, and then the process is iterated until the poster
distribution approximation appears to converge wrt. Hellinger distance.
}
\details{
The method is experimental because its theoretical convergence properties 
are not fully explored.

# @example examples/stochit.R
}
