% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsdive.suffstat.R
\name{dsdive.suffstat}
\alias{dsdive.suffstat}
\title{Compute sufficient statistics for posterior distributions of complete dives}
\usage{
dsdive.suffstat(depths, times, t.stages, depth.bins)
}
\description{
Compute sufficient statistics for posterior distributions of complete dives
}
\examples{
data('dive.sim')
attach(dive.sim)
attach(dive.sim$params)

# get true stage transition times
t.stages = sim$times[c(FALSE, diff(sim$stages)==1)]

dsdive.suffstat(depths = sim$depths, times = sim$times, t.stages = t.stages, 
                depth.bins = depth.bins)

detach(dive.sim$params)
detach(dive.sim)
}
