% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsdive.tx.params.R
\name{dsdive.tx.params}
\alias{dsdive.tx.params}
\title{Compute transition parameters for dive trajectories across discrete depths}
\usage{
dsdive.tx.params(t0, num.depths, d0, d0.last = NULL, s0, beta, lambda,
  sub.tx, surf.tx)
}
\arguments{
\item{t0}{time at which transition parameters should be computed}

\item{num.depths}{The total number of depth bins.  Specifies the 
depth bins 1, ..., \code{num.depths} the model will use.  Note that the 
model does not explicitly use the depth ranges the bins represent 
(i.e., the bin labels) to compute transition parameters.}

\item{d0}{the depth bin at which transition parameters should be computed}

\item{d0.last}{the previous depth bin in which the trajectory was.  If 
\code{NULL}, then the autoregressive component will be skipped.}

\item{s0}{the dive stage (PRIMARY DESCENT==1, INTERMEDIATE BEHAVIORS==2, 
PRIMARY ASCENT==3) for which transition parameters should be computed.}

\item{beta}{\eqn{2 x 3} matrix in which each column contains the diving 
preference and directional persistence parameters for the PRIMARY DESCENT 
(PD), INTERMEDIATE BEHAVIORS (IB), and PRIMARY ASCENT (PA) dive stages.}

\item{lambda}{length 3 vector that specifies the transition rate, 
respectively in the PD, IB, and PA stages.}

\item{sub.tx}{length 2 vector that specifies the first depth bin at which 
transitions to the IB stage can occur and the probability that such 
a transition occurs at the next depth transition}

\item{surf.tx}{parameter that specifies the probability the trajectory will 
transition to the PA stage at the next depth transition}
}
\description{
Computes elements of a transition matrix when given model parameters and 
time/space locations for a dive model that has 3 stages, PRIMARY DESCENT 
(PD), INTERMEDIATE BEHAVIORS (IB), and PRIMARY ASCENT (PA).
}
\examples{
data('dive.sim')
attach(dive.sim)
attach(dive.sim$params)

dsdive.tx.params(t0 = 0, num.depths = nrow(depth.bins), d0 = 0, 
                 d0.last = NULL, s0 = 1, beta = beta, lambda = lambda, 
                 sub.tx = sub.tx, surf.tx = surf.tx)

dsdive.tx.params(t0 = 0, num.depths = nrow(depth.bins), d0 = 60, 
                 d0.last = 61, s0 = 1, beta = beta, lambda = lambda, 
                 sub.tx = sub.tx, surf.tx = surf.tx)

dsdive.tx.params(t0 = 0, num.depths = nrow(depth.bins), d0 = 60, 
                 d0.last = 61, s0 = 2, beta = beta, lambda = lambda, 
                 sub.tx = sub.tx, surf.tx = surf.tx)
}
