% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsdive.tx.params.R
\name{dsdive.tx.params}
\alias{dsdive.tx.params}
\title{Compute transition parameters for dive trajectories across discrete depths}
\usage{
dsdive.tx.params(depth.bins, d0, s0, beta, lambda)
}
\arguments{
\item{depth.bins}{\eqn{n x 2} Matrix that defines the depth bins.  The first 
column defines the depth at the center of each depth bin, and the second 
column defines the half-width of each bin.}

\item{d0}{the depth bin at which transition parameters should be computed}

\item{s0}{the dive stage (PRIMARY DESCENT==1, INTERMEDIATE BEHAVIORS==2, 
PRIMARY ASCENT==3) for which transition parameters should be computed.}

\item{beta}{\eqn{2 x 3} matrix in which each column contains the diving 
preference and directional persistence parameters for the PRIMARY DESCENT 
(PD), INTERMEDIATE BEHAVIORS (IB), and PRIMARY ASCENT (PA) dive stages.}

\item{lambda}{length 3 vector that specifies the transition rate, 
respectively in the PD, IB, and PA stages.}

\item{t0}{time at which transition parameters should be computed}

\item{d0.last}{the previous depth bin in which the trajectory was.  If 
\code{NULL}, then the autoregressive component will be skipped.}

\item{sub.tx}{length 2 vector that specifies the first depth bin at which 
transitions to the IB stage can occur and the probability that such 
a transition occurs at the next depth transition}

\item{surf.tx}{parameter that specifies the probability the trajectory will 
transition to the PA stage at the next depth transition}

\item{t0.dive}{Time at which dive started}

\item{shift.params}{If not \code{NULL}, then will apply a directional bias 
to depth bin transition probabilities.  \code{shift.params} should be a
vector \code{c(df, p.bias, tf, lambda.bias)}. The first element is the 
target depth bin to travel to.
The second element controls the strength of the bias.  The second element 
should be a positive value, with 1.125 being the recommended value.  Using 
\code{shift.params[2] = 1.125} will reinforce natural movement in good 
directions of travel, and replace "poor" directions of travel with bias in 
"good" directions of travel.  Setting \code{shift.params[2] = 1} will 
replace "poor" directions of travel with random walk proposals.  Values 
less than 1 will push poor directions of travel closer toward random walks
from their natural tendencies.}

\item{t.stage2}{time at which second stage was entered.}

\item{model}{Either \code{"conditional"} or \code{"logit"} depending on the 
method used to determine stage transition probability curves}
}
\description{
Computes elements of a transition matrix when given model parameters and 
time/space locations for a dive model that has 3 stages, PRIMARY DESCENT 
(PD), INTERMEDIATE BEHAVIORS (IB), and PRIMARY ASCENT (PA).
}
\examples{
data('dive.sim')
attach(dive.sim)
attach(params)

dsdive.tx.params(depth.bins = depth.bins, d0 = 3, s0 = 1, 
                 beta = beta, lambda = lambda)

detach(params)
detach(dive.sim)
}
