% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envelope.logquad.R
\name{envelope.logquad}
\alias{envelope.logquad}
\title{Build a piecewise-quadratic envelope given log of a function}
\usage{
envelope.logquad(breaks, logf, d.logf, dd.logf.sup)
}
\arguments{
\item{breaks}{Breakpoints that define the start and end points of each 
interval over which a piecewise-quadratic function will be defined.}

\item{f}{Value of f at the breakpoints.}

\item{df}{Derivative of f evaluated at the start of each interval.}

\item{ddf.sup}{Supremum of second derivative of f over each interval}
}
\description{
Build a piecewise-quadratic envelope given log of a function
}
\examples{
library(truncdist)

# define range for the envelope
xmin = 1
xmax = 10

# define envelope segments
breaks = seq(from = xmin, to = xmax, length.out = 10)

# get number of segments
n = length(breaks)


#
# example 1: f(x) = exp(-x); log f(x) = -x
#

devtools::document()

# build envelope
q = envelope.logquad(breaks = breaks, 
                     logf = -breaks[1:(n-1)], 
                     d.logf = rep(-1, n-1), 
                     dd.logf.sup = rep(0, n-1))

# sample from density
samples = q$rquad(n = 1e4)

# compare density functions, and kernel density estimate
curve(dtrunc(x = x, spec = 'exp', a = xmin, b = xmax), from = xmin, to = xmax, 
      ylab = expression(f(x)))
curve(q$dquad(x), from = xmin, to = xmax, col = 2, lty = 2, add = TRUE)
lines(density(samples), col = 3, lty = 2)


#
# example 2: f(x) = exp(sin(x)); log f(x) = sin(x)
#


# build envelope
q = envelope.logquad(breaks = breaks, 
                     logf = sin(breaks[1:(n-1)]), 
                     d.logf = cos(breaks[1:(n-1)]), 
                     dd.logf.sup = sapply(1:(n-1), function(i) {
                       max(-sin(c(breaks[i:(i+1)])))
                     }))

# look at envelope
curve(exp(sin(x)), from = xmin, to = xmax, 
      ylab = expression(f(x)))
curve(q$e(x), from = xmin, to = xmax, col = 2, lty = 2, add = TRUE)


# sample from density
samples = q$rquad(n = 1e4)

# compare density to kernel density estimate
curve(q$dquad(x), from = xmin, to = xmax, ylab = expression(f(x)))
lines(density(samples), col = 3, lty = 2)
}
