% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envelope.quad.R
\name{envelope.quad}
\alias{envelope.quad}
\title{Build a piecewise-quadratic envelope for a given function}
\usage{
envelope.quad(breaks, f, df, ddf.sup, anchors = breaks[1:(length(breaks) - 1)])
}
\arguments{
\item{breaks}{Breakpoints that define the start and end points of each 
interval over which a piecewise-quadratic function will be defined.}

\item{f}{Value of f at the anchors.}

\item{df}{Derivative of f evaluated at the anchors.}

\item{ddf.sup}{Supremum of second derivative of f over each interval}

\item{anchors}{Locations within the intervals (defined by \code{breaks}) at 
which the polynomial approximation will be centered}
}
\description{
Build a piecewise-quadratic envelope for a given function
}
