% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envelope.quad.R
\name{envelope.quad}
\alias{envelope.quad}
\title{Build a piecewise-quadratic envelope for a given function}
\usage{
envelope.quad(breaks, f, df, ddf.sup)
}
\arguments{
\item{breaks}{Breakpoints that define the start and end points of each 
interval over which a piecewise-quadratic function will be defined.}

\item{f}{Value of f at the breakpoints.}

\item{df}{Derivative of f evaluated at the start of each interval.}

\item{ddf.sup}{Supremum of second derivative of f over each interval}
}
\description{
Build a piecewise-quadratic envelope for a given function
}
\examples{
# define range for the envelope
xmin = 1
xmax = 10

# define envelope segments
breaks = seq(from = xmin, to = xmax, length.out = 10)

# get number of segments
n = length(breaks)


#
# example 1: f(x) = exp(x)
#

# build envelope
q = envelope.quad(breaks = breaks, 
                  f = exp(breaks[1:(n-1)]), 
                  df = exp(breaks[1:(n-1)]), 
                  ddf.sup = exp(breaks[-1]))

# compare envelope to target function
curve(exp(x), from = xmin, to = xmax)
curve(q(x), from = xmin, to = xmax, add = TRUE, col = 4, n = 1e3)


#
# example 2: f(x) = exp(-x)
#

# build envelope
q2 = envelope.quad(breaks = breaks, 
                   f = exp(-breaks[1:(n-1)]), 
                   df = -exp(-breaks[1:(n-1)]), 
                   ddf.sup = exp(-breaks[1:(n-1)]))

# compare envelope to target function
curve(exp(-x), from = xmin, to = xmax)
curve(q2(x), from = xmin, to = xmax, add = TRUE, col = 4, n = 1e3)
}
