% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussapprox.R
\name{gaussapprox}
\alias{gaussapprox}
\title{Compute a Gaussian approximation to a density}
\usage{
gaussapprox(
  logf,
  init,
  method = "BFGS",
  optim.output = FALSE,
  lower = -Inf,
  upper = Inf,
  control = list(fnscale = -1)
)
}
\arguments{
\item{logf}{log of the density for which an approximation should be built.}

\item{init}{initial guess for mode of \code{logf}}

\item{method}{method to be used in \code{stats::optim}}

\item{optim.output}{\code{TRUE} to return output from \code{stats::optim}}

\item{lower}{lower bound value to pass to \code{stats::optim}}

\item{upper}{upper bound value to pass to \code{stats::optim}}

\item{control}{control list to pass to \code{stats::optim}}
}
\description{
Computes a Gaussian approximation to the target density \code{logf}, and 
additionally return methods to aid in sampling from the approximation.  The 
approximation is computed numerically via the \code{stats::optim} routine.
}
