% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.convert.R
\name{params.toList}
\alias{params.toList}
\title{Convert vector of parameters on transformed scale to a list of parameters on model scale}
\usage{
params.toList(par, sub.tx1)
}
\arguments{
\item{par}{the vector of parameters to convert to a list of model parameters}

\item{sub.tx1}{the fixed value of sub.tx1 to incorporate into the parameter 
list}
}
\description{
The output also includes the log-Jacobian for the transformation
}
\examples{

data('dive.sim')
attach(dive.sim)
attach(params)

params.toList(par = runif(12), sub.tx1 = 1)

detach(params)
detach(dive.sim)

}
