% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.convert.R
\name{params.toList}
\alias{params.toList}
\title{Convert vector of parameters on transformed scale to a list of parameters on model scale}
\usage{
params.toList(par, spec)
}
\arguments{
\item{par}{the vector of parameters to convert to a list of model parameters}

\item{spec}{List of parameters to specify prior distributions.  See 
\code{dsdive.prior} for more details.}
}
\description{
The output also includes the log-Jacobian for the transformation.  Note that 
the transformation (and associated Jacobian) only links the unrestricted 
parameter space over which the Gibbs sampler is defined to the parameter 
space used by the model.  If prior distributions are defined for additional 
transformations of the model parameters, than the additional Jacobians must 
be accounted for within the \code{dsdive.prior} function.
}
\examples{
data('dive.sim')
attach(dive.sim)
attach(params)

spec = list(beta.sd = 1, beta.absmax = 10, lambda.sd = 1, sub.tx.mean = 0, 
            sub.tx.sd = 1, surf.tx.mean = 0, surf.tx.sd = 1)

v = params.toVec(par = params, spec = spec)

L = params.toList(par = v)

detach(params)
detach(dive.sim)
}
