% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.convert.R
\name{params.toVec}
\alias{params.toVec}
\title{Convert list of parameters to vector of parameters on transformed}
\usage{
params.toVec(par)
}
\arguments{
\item{par}{the list of parameters to convert to a vector of model parameters}
}
\description{
Convert list of parameters to vector of parameters on transformed
}
\examples{

data('dive.sim')
attach(dive.sim)
attach(params)

params.toVec(par = params)

detach(params)
detach(dive.sim)

}
