% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.dsdive.R
\name{plot.dsdive}
\alias{plot.dsdive}
\title{Plotting functions for completely observed dive trajectories}
\usage{
plot.dsdive(
  x,
  depth.bins,
  imputed.alpha = 0.3,
  dsobs = NULL,
  imputed.list = NULL,
  underlay = NULL,
  time.as.POSIXct = FALSE,
  ...
)
}
\arguments{
\item{x}{\code{dsdive} object with complete record of dive trajectory}

\item{depth.bins}{\eqn{n x 2} Matrix that defines the depth bins.  The first 
column defines the depth at the center of each depth bin, and the second 
column defines the half-width of each bin.}

\item{imputed.alpha}{transparency value used for plotting imputed 
trajectories, if provided}

\item{dsobs}{\code{dsobs} object with record of observed depth bins to 
overlay on the main \code{dsdive} trajectory \code{x}}

\item{imputed.list}{A \code{list} of \code{dsdive} objects to be plotted 
underneath the main \code{dsdive} trajectory \code{x}.  The intent is that 
\code{imputed.list} will contain imputed trajectories.}

\item{underlay}{additional \code{ggplot2} layers to place before the main 
plot layers}

\item{time.as.POSIXct}{if \code{TRUE}, will convert plotting times to POSIXct}

\item{...}{(currently unused) additional plotting parameters}
}
\description{
Builds a \code{ggplot2} graph.
}
\examples{
data('dive.sim')
attach(dive.sim)

# basic plot
pl = plot(x = sim, depth.bins = depth.bins)

# plot with observations overlaid
pl2 = plot(x = sim, depth.bins = depth.bins, dsobs = sim.obs)

# plot with "imputed" trajectory overlaid
pl3 = plot(x = sim, depth.bins = depth.bins, dsobs = sim.obs, 
           imputed.list = sim)

detach(dive.sim)
}
