% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.dsobs.R
\name{plot.dsobs}
\alias{plot.dsobs}
\title{Plotting functions for observations of dive trajectories}
\usage{
plot.dsobs(
  x,
  depth.bins,
  stages = NULL,
  errorbars = FALSE,
  underlay = NULL,
  imputed.alpha = 0.3,
  imputed.list = NULL,
  time.as.POSIXct = FALSE,
  ...
)
}
\arguments{
\item{x}{\code{dsobs} object with record of observed depth bins}

\item{depth.bins}{\eqn{n x 2} Matrix that defines the depth bins.  The first 
column defines the depth at the center of each depth bin, and the second 
column defines the half-width of each bin.}

\item{stages}{Optional vector of for which dive stage the trajectory was 
in at each observation}

\item{errorbars}{If \code{TRUE}, then the minimum and maximum depth for each
bin will be included in the plot.}

\item{underlay}{additional \code{ggplot2} layers to place before the main 
plot layers}

\item{imputed.alpha}{transparency value used for plotting imputed 
trajectories, if provided}

\item{imputed.list}{A \code{list} of \code{dsdive} objects to be plotted 
underneath the main \code{dsdive} trajectory \code{x}.  The intent is that 
\code{imputed.list} will contain imputed trajectories.}

\item{time.as.POSIXct}{if \code{TRUE}, will convert plotting times to POSIXct}

\item{...}{(currently unused) additional plotting parameters}
}
\description{
Builds a \code{ggplot2} graph.
}
\examples{
data('dive.sim')
attach(dive.sim)

# basic plot
pl = plot(x = sim.obs, depth.bins = depth.bins)

# plot with stages overlaid
pl2 = plot(x = sim.obs, depth.bins = depth.bins, stages = sim.obs$stages)

# plot with trajectory underlaid
pl3 = plot(x = sim.obs, depth.bins = depth.bins, imputed.list = sim)

detach(dive.sim)
}
