% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyval.R
\name{polyval}
\alias{polyval}
\title{Use Horner's method to efficiently evaluate an n degree polynomial}
\usage{
polyval(coefs, x)
}
\arguments{
\item{coefs}{polynomial coefficients in decreasing order}

\item{x}{vector of values of x at which to evaluate polynomial

# @example examples/polyval.R}
}
\description{
f(x) = a_n * x^n + a_{n-1} * x^{n-1} + ... + a_0
}
