% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refine.partition.R
\name{refine.partition}
\alias{refine.partition}
\title{Refine a partition of an interval}
\usage{
refine.partition(breaks, max.width)
}
\description{
The input \code{breaks} defines a partition.  \code{refine.partition} will 
add evenly spaced points between \code{breaks} such that the maximum 
difference between any two points in the output is \code{max.width}.
}
\examples{
data('dive.sim')
attach(dive.sim)

# build partition of sim$times that 1) includes sim$times, and 2) has a 
# maximum increment of 100
times.refined = refine.partition(breaks = sim$times, max.width = 100)

# find where the original times appear in the refined partition
refined.mapping = sapply(sim$times, function(x) which(x==times.refined))

# display the difference in the partition
times.compare = rep(NA, length(times.refined))
times.compare[refined.mapping] = sim$times
cbind(original = times.compare, 
      refined = times.refined, 
      width = c(diff(times.refined), NA))

detach(dive.sim)
}
