% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spsa.R
\name{spsa}
\alias{spsa}
\title{Basic implementation of stochastic optimization when the target function is only approximately known}
\usage{
spsa(par, fn, maxit = 100, a = 1, c = 1, A = 1, alpha = 0.5,
  gamma = 1, tol = 0.001)
}
\arguments{
\item{par}{initial parameter values for function}

\item{fn}{a function to be maximized}

\item{maxit}{maximum number of function evaluations before failing to converge}

\item{a}{decay parameter for step size}

\item{c}{decay parameter for step size}

\item{A}{decay parameter for step size}

\item{alpha}{decay parameter for step size}

\item{gamma}{decay parameter for step size}

\item{tol}{Relative convergence criterion}
}
\description{
Experimental method because its theoretical convergence and recommendations 
for setting parameters has not been fully explored.
}
\examples{
data('dive.sim')
attach(dive.sim)
attach(dive.sim$params)

o = spsa(par = c(beta, log(lambda), log(sub.tx[-1]), surf.tx), 
          fn = function(theta) {
            r = dsdive.ldabc(beta = matrix(theta[1:6], nrow = 2), 
                             lambda = exp(theta[7:9]), 
                             sub.tx = c(sub.tx[1], exp(theta[10])), 
                             surf.tx = theta[11:12], 
                             depth.bins = depth.bins, 
                             steps.max = 1e3, N = 1e1, depths = sim.obs$depths, 
                             t = sim.obs$times, tries.max = 1e5, 
                             dump.state = TRUE, verbose = FALSE, n.samples = 1, 
                             eps = 1)$ld
            -r
          }, c = 2^(-6), a = 2^(-12), alpha = 1/3)
}
