% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stagevec.R
\name{stagevec}
\alias{stagevec}
\title{Convenience function for building a vector for dive stages}
\usage{
stagevec(length.out, breaks)
}
\arguments{
\item{length.out}{Length of vector to generate}

\item{breaks}{Location of changepoints}
}
\description{
Convenience function for building a vector for dive stages
}
\examples{

stagevec(length.out = 30, breaks = c(7, 20))

}
